Module tabled::grid

Module is responsible for tables underlyign grid.

It might be used when implementing your own TableOption and CellOption.

Modules

color

A module which contains Color trait and its implementation AnsiColor.

colors

A module which contains Colors trait and its blanket implementations.

config

Module contains a list of configs for varios tables/grids.

dimension

Module contains a list of implementations of Estimate and Dimension.

records

The module contains Records, ExactRecords, RecordsMut, Resizable traits and its implementations.

util

A module contains utility functions which grid relay on.

Structs

CompactGrid

Grid provides a set of methods for building a text-based table.

Grid

Grid provides a set of methods for building a text-based table.

PeekableGrid

Grid provides a set of methods for building a text-based table.