tabled::grid

Module config

Module contains a list of configs for varios tables/grids.

Structs

Border

Border is a representation of a cells’s borders (left, right, top, bottom, and the corners)

Borders

Borders represents a Table frame with horizontal and vertical split lines.

ColorMap

A colors structure for ColoredConfig.

ColoredConfig

A spanned configuration plus colors for cells.

CompactConfig

This structure represents a settings of a grid.

CompactMultilineConfig

A CompactConfig based configuration plus vertical alignment and formatting options.

EntityIterator

An iterator over cells.

EntityMap

A structure to keep information for Entity as a key.

Formatting

Formatting represent a logic of formatting of a cell text.

HorizontalLine

A structure for a custom horizontal line.

Indent

Indent represent a filled space.

Position

Position is a (row, col) position on a Grid.

Sides

A structure which represents 4 box sides.

SpannedConfig

This structure represents a settings of a grid.

VerticalLine

A structure for a vertical line.

Enums

AlignmentHorizontal

AlignmentHorizontal represents an horizontal alignment of a cell content.

AlignmentVertical

AlignmentVertical represents an vertical alignment of a cell content.

Entity

Entity a structure which represent a set of cells.

Offset

The structure represents an offset in a text.