Struct tabled::grid::config::SpannedConfig

pub struct SpannedConfig { /* private fields */ }

This structure represents a settings of a grid.

grid: crate::Grid.

Implementations

impl SpannedConfig

pub fn set_margin(&mut self, margin: Sides<Indent>)

Set a margin of a grid.

pub fn set_margin_color(&mut self, margin: Sides<Option<AnsiColor<'static>>>)

Set a color of margin of a grid.

pub fn set_margin_offset(&mut self, margin: Sides<Offset>)

Set an offset of margin of a grid.

pub fn get_margin(&self) -> Sides<Indent>

Returns a margin value currently set.

pub fn get_margin_color(&self) -> Sides<Option<AnsiColor<'static>>>

Returns a margin color value currently set.

pub fn get_margin_offset(&self) -> Sides<Offset>

Returns a margin offset value currently set.

pub fn clear_theme(&mut self)

Clears all theme changes. And sets it to default.

pub fn set_borders(&mut self, borders: Borders<char>)

Set the Borders value as correct one.

pub fn get_global_border(&self) -> Option<&char>

Gets a global border value if set.

pub fn set_global_border(&mut self, c: char)

Set the all Borders values to a char.

pub fn get_borders(&self) -> &Borders<char>

Returns a current Borders structure.

pub fn insert_horizontal_line(&mut self, line: usize, val: HorizontalLine<char>)

Set the border line by row index.

Row 0 means the top row. Row grid.count_rows() means the bottom row.

pub fn remove_horizontal_line(&mut self, line: usize, count_rows: usize)

Sets off the border line by row index if any were set

Row 0 means the top row. Row grid.count_rows() means the bottom row.

pub fn get_vertical_line(&self, line: usize) -> Option<&VerticalLine<char>>

Gets a overridden vertical line.

Row 0 means the left row. Row grid.count_columns() means the right most row.

pub fn insert_vertical_line(&mut self, line: usize, val: VerticalLine<char>)

Set the border line by column index.

Row 0 means the left row. Row grid.count_columns() means the right most row.

pub fn remove_vertical_line(&mut self, line: usize, count_columns: usize)

Sets off the border line by column index if any were set

Row 0 means the left row. Row grid.count_columns() means the right most row.

pub fn get_horizontal_line(&self, line: usize) -> Option<&HorizontalLine<char>>

Gets a overridden line.

Row 0 means the top row. Row grid.count_rows() means the bottom row.

pub fn set_horizontal_char( &mut self, pos: (usize, usize), c: char, offset: Offset )

Override a character on a horizontal line.

If borders are not set the char won’t be used.

It takes not cell position but line as row and column of a cell; So its range is line <= count_rows && col < count_columns.

pub fn lookup_horizontal_char( &self, pos: (usize, usize), offset: usize, end: usize ) -> Option<char>

Get a list of overridden chars in a horizontal border.

It takes not cell position but line as row and column of a cell; So its range is line <= count_rows && col < count_columns.

pub fn is_overridden_horizontal(&self, pos: (usize, usize)) -> bool

Checks if there any char in a horizontal border being overridden.

It takes not cell position but line as row and column of a cell; So its range is line <= count_rows && col < count_columns.

pub fn remove_overridden_horizontal(&mut self, pos: (usize, usize))

Removes a list of overridden chars in a horizontal border.

It takes not cell position but line as row and column of a cell; So its range is line <= count_rows && col < count_columns.

pub fn set_vertical_char( &mut self, pos: (usize, usize), c: char, offset: Offset )

Override a vertical split line.

If borders are not set the char won’t be used.

It takes not cell position but cell row and column of a line; So its range is row < count_rows && col <= count_columns.

pub fn lookup_vertical_char( &self, pos: (usize, usize), offset: usize, end: usize ) -> Option<char>

Get a list of overridden chars in a horizontal border.

It takes not cell position but cell row and column of a line; So its range is row < count_rows && col <= count_columns.

pub fn is_overridden_vertical(&self, pos: (usize, usize)) -> bool

Checks if there any char in a horizontal border being overridden.

It takes not cell position but cell row and column of a line; So its range is row < count_rows && col <= count_columns.

pub fn remove_overridden_vertical(&mut self, pos: (usize, usize))

Removes a list of overridden chars in a horizontal border.

It takes not cell position but cell row and column of a line; So its range is row < count_rows && col <= count_columns.

pub fn set_horizontal_color( &mut self, pos: (usize, usize), c: AnsiColor<'static>, offset: Offset )

Override a character color on a horizontal line.

pub fn lookup_horizontal_color( &self, pos: (usize, usize), offset: usize, end: usize ) -> Option<&AnsiColor<'static>>

Get a overridden color in a horizontal border.

pub fn set_vertical_color( &mut self, pos: (usize, usize), c: AnsiColor<'static>, offset: Offset )

Override a character color on a vertical line.

pub fn lookup_vertical_color( &self, pos: (usize, usize), offset: usize, end: usize ) -> Option<&AnsiColor<'static>>

Get a overridden color in a vertical border.

pub fn set_padding(&mut self, entity: Entity, padding: Sides<Indent>)

Set a padding to a given cells.

pub fn set_padding_color( &mut self, entity: Entity, padding: Sides<Option<AnsiColor<'static>>> )

Set a padding to a given cells.

pub fn get_padding(&self, entity: Entity) -> Sides<Indent>

Get a padding for a given Entity.

pub fn get_padding_color( &self, entity: Entity ) -> Sides<Option<AnsiColor<'static>>>

Get a padding color for a given Entity.

pub fn set_formatting(&mut self, entity: Entity, formatting: Formatting)

Set a formatting to a given cells.

pub fn get_formatting(&self, entity: Entity) -> &Formatting

Get a formatting settings for a given Entity.

pub fn set_alignment_vertical( &mut self, entity: Entity, alignment: AlignmentVertical )

Set a vertical alignment to a given cells.

pub fn get_alignment_vertical(&self, entity: Entity) -> &AlignmentVertical

Get a vertical alignment for a given Entity.

pub fn set_alignment_horizontal( &mut self, entity: Entity, alignment: AlignmentHorizontal )

Set a horizontal alignment to a given cells.

pub fn get_alignment_horizontal(&self, entity: Entity) -> &AlignmentHorizontal

Get a horizontal alignment for a given Entity.

pub fn set_border(&mut self, pos: (usize, usize), border: Border<char>)

Set border set a border value to all cells in Entity.

pub fn get_border( &self, pos: (usize, usize), shape: (usize, usize) ) -> Border<char>

Returns a border of a cell.

pub fn get_border_color( &self, pos: (usize, usize), shape: (usize, usize) ) -> Border<&AnsiColor<'static>>

Returns a border color of a cell.

pub fn set_borders_missing(&mut self, c: char)

Set a character which will be used in case any misconfiguration of borders.

It will be usde for example when you set a left char for border frame and top but didn’t set a top left corner.

pub fn get_borders_missing(&self) -> char

Get a character which will be used in case any misconfiguration of borders.

pub fn get_border_color_global(&self) -> Option<&AnsiColor<'static>>

Gets a color of all borders on the grid.

pub fn set_border_color_global(&mut self, clr: AnsiColor<'static>)

Sets a color of all borders on the grid.

pub fn get_color_borders(&self) -> &Borders<AnsiColor<'static>>

Gets colors of a borders carcass on the grid.

pub fn set_borders_color(&mut self, clrs: Borders<AnsiColor<'static>>)

Sets colors of border carcass on the grid.

pub fn set_border_color( &mut self, pos: (usize, usize), border: Border<AnsiColor<'static>> )

Sets a color of border of a cell on the grid.

pub fn remove_border(&mut self, pos: (usize, usize), shape: (usize, usize))

Sets off all borders possible on the Entity.

It doesn’t changes globally set borders through SpannedConfig::set_borders.

pub fn remove_border_color( &mut self, pos: (usize, usize), shape: (usize, usize) )

Gets a color of border of a cell on the grid.

pub fn get_justification(&self, entity: Entity) -> char

Get a justification which will be used while expanding cells width/height.

pub fn get_justification_color( &self, entity: Entity ) -> Option<&AnsiColor<'static>>

Get a justification color which will be used while expanding cells width/height.

None means no color.

pub fn set_justification(&mut self, entity: Entity, c: char)

Set a justification which will be used while expanding cells width/height.

pub fn set_justification_color( &mut self, entity: Entity, color: Option<AnsiColor<'static>> )

Set a justification color which will be used while expanding cells width/height.

None removes it.

pub fn get_column_spans(&self) -> HashMap<(usize, usize), usize, RandomState>

Get a span value of the cell, if any is set.

pub fn get_row_spans(&self) -> HashMap<(usize, usize), usize, RandomState>

Get a span value of the cell, if any is set.

pub fn get_column_span(&self, pos: (usize, usize)) -> Option<usize>

Get a span value of the cell, if any is set.

pub fn get_row_span(&self, pos: (usize, usize)) -> Option<usize>

Get a span value of the cell, if any is set.

pub fn remove_column_spans(&mut self)

Removes column spans.

pub fn remove_row_spans(&mut self)

Removes row spans.

pub fn set_column_span(&mut self, pos: (usize, usize), span: usize)

Set a column span to a given cells.

BEWARE

IT’S CALLER RESPONSIBILITY TO MAKE SURE THAT THERE NO INTERSECTIONS IN PLACE AND THE SPAN VALUE IS CORRECT

pub fn has_column_spans(&self) -> bool

Verifies if there’s any spans set.

pub fn set_row_span(&mut self, pos: (usize, usize), span: usize)

Set a column span to a given cells.

BEWARE

IT’S CALLER RESPONSIBILITY TO MAKE SURE THAT THERE NO INTERSECTIONS IN PLACE AND THE SPAN VALUE IS CORRECT

pub fn has_row_spans(&self) -> bool

Verifies if there’s any spans set.

pub fn get_intersection( &self, pos: (usize, usize), shape: (usize, usize) ) -> Option<char>

Gets an intersection character which would be rendered on the grid.

grid: crate::Grid

pub fn get_horizontal( &self, pos: (usize, usize), count_rows: usize ) -> Option<char>

Gets a horizontal character which would be rendered on the grid.

grid: crate::Grid

pub fn get_vertical( &self, pos: (usize, usize), count_columns: usize ) -> Option<char>

Gets a vertical character which would be rendered on the grid.

grid: crate::Grid

pub fn get_horizontal_color( &self, pos: (usize, usize), count_rows: usize ) -> Option<&AnsiColor<'static>>

Gets a color of a cell horizontal.

pub fn get_vertical_color( &self, pos: (usize, usize), count_columns: usize ) -> Option<&AnsiColor<'static>>

Gets a color of a cell vertical.

pub fn get_intersection_color( &self, pos: (usize, usize), shape: (usize, usize) ) -> Option<&AnsiColor<'static>>

Gets a color of a cell vertical.

pub fn has_horizontal(&self, row: usize, count_rows: usize) -> bool

Checks if grid would have a horizontal border with the current configuration.

grid: crate::Grid

pub fn has_vertical(&self, col: usize, count_columns: usize) -> bool

Checks if grid would have a vertical border with the current configuration.

grid: crate::Grid

pub fn count_horizontal(&self, count_rows: usize) -> usize

Calculates an amount of horizontal lines would present on the grid.

grid: crate::Grid

pub fn count_vertical(&self, count_columns: usize) -> usize

Calculates an amount of vertical lines would present on the grid.

grid: crate::Grid

pub fn is_cell_visible(&self, pos: (usize, usize)) -> bool

The function returns whether the cells will be rendered or it will be hidden because of a span.

pub fn is_cell_covered_by_row_span(&self, pos: (usize, usize)) -> bool

The function checks if a cell is hidden because of a row span.

pub fn is_cell_covered_by_column_span(&self, pos: (usize, usize)) -> bool

The function checks if a cell is hidden because of a column span.

pub fn is_cell_covered_by_both_spans(&self, pos: (usize, usize)) -> bool

The function checks if a cell is hidden indirectly because of a row and column span combination.