tabled::grid

Module records

The module contains Records, ExactRecords, RecordsMut, Resizable traits and its implementations.

Also it provides a list of helpers for a user built Records via into_records.

Modules

into_records

The module contains a list of helpers for IntoRecords

vec_records

Module contains VecRecords.

Structs

EmptyRecords

Empty representation of Records.

IterRecords

A Records implementation for any IntoIterator.

Traits

ExactRecords

Records extension which guarantees the amount of rows.

IntoRecords

The representation of data, rows and columns of a grid.

PeekableRecords

The representation of data, rows and columns of a grid.

Records

Records represents table data.

RecordsMut

A Records representation which can modify cell by (row, column) index.

Resizable

A records representation which can be modified by moving rows/columns around.