tabled::grid::records

Trait PeekableRecords

pub trait PeekableRecords {
    // Required method
    fn get_text(&self, pos: Position) -> &str;

    // Provided methods
    fn get_line(&self, pos: Position, line: usize) -> &str { ... }
    fn count_lines(&self, pos: Position) -> usize { ... }
    fn get_width(&self, pos: Position) -> usize { ... }
    fn get_line_width(&self, pos: Position, line: usize) -> usize { ... }
}

The representation of data, rows and columns of a grid.

Required Methods

fn get_text(&self, pos: Position) -> &str

Returns a text of a cell by an index.

Provided Methods

fn get_line(&self, pos: Position, line: usize) -> &str

Returns a line of a text of a cell by an index.

fn count_lines(&self, pos: Position) -> usize

Returns an amount of lines of a text of a cell by an index.

fn get_width(&self, pos: Position) -> usize

Returns a width of a text of a cell by an index.

fn get_line_width(&self, pos: Position, line: usize) -> usize

Returns a width of line of a text of a cell by an index.