tabled::grid::records

Trait RecordsMut

pub trait RecordsMut<Text> {
    // Required method
    fn set(&mut self, pos: Position, text: Text);
}

A Records representation which can modify cell by (row, column) index.

Required Methods

fn set(&mut self, pos: Position, text: Text)

Sets a text to a given cell by index.