tabled::grid::records

Module vec_records

Module contains VecRecords.

Structs

StrWithWidth

StrWithWidth is a structure is responsible for a string and it’s width.

Text

The struct is a Cell implementation which keeps width information pre allocated.

VecRecords

A Records implementation based on allocated buffers.

Traits

Cell

Cell implementation which can be used with VecRecords.