tabled::grid::util

Module string

This module contains a different functions which are used by the IterGrid.

You should use it if you want to comply with how IterGrid.

Structs

Lines

Iterator over lines.

Functions

count_lines

Calculates a number of lines.

count_tabs

Returns a list of tabs (\t) in a string..

get_char_width

Returns a char width.

get_line_width

Returns a string width.

get_lines

Splits the string by lines.

get_string_width

Returns a string width (accouting all characters).

get_text_dimension

Returns string width and count lines of a string. It’s a combination of get_text_width and count_lines.

get_text_width

Returns a max string width of a line.

replace_tab

Replaces tabs in a string with a given width of spaces.