Trait tabled::locator::Locator

pub trait Locator {
    type Coordinate;
    type IntoIter: IntoIterator<Item = Self::Coordinate>;

    // Required method
    fn locate<R>(&mut self, records: R) -> Self::IntoIter
       where R: Records;
}

Locator is an interface which searches for a particular thing in the Records, and returns coordinate of the foundings if any.

Required Associated Types

type Coordinate

A coordinate of the finding.

type IntoIter: IntoIterator<Item = Self::Coordinate>

An iterator of the coordinates. If it’s empty it’s consideret that nothing is found.

Required Methods

fn locate<R>(&mut self, records: R) -> Self::IntoIterwhere R: Records,

Search for the thing in Records, returning a list of coordinates.