Module tabled::object

This module contains a list of primitives that implement a Object trait. They help to locate a nessesary segment on a Table.

Structs

Cell

Cell denotes a particular cell on a Table.

Column

Column represents a single column on a grid.

Columns

Column denotes a set of cells on given columns on a Table.

Combination

Combination struct used for chaning Object’s.

FirstColumn

FirstColumn represents the first column on a grid.

FirstRow

FirstRow represents the first row of a Table. It’s often contains headers data.

Frame

Frame includes cells which are on the edges of each side. Therefore it’s Object implementation returns a subset of cells which are present in frame.

FullDeprecated

Full represents all cells on a Table.

LastColumn

LastColumn represents the last column on a grid.

LastColumnOffset

LastColumnOffset represents a single column on a grid indexed via offset from the last column.

LastRow

LastRow represents the last row of a Table.

LastRowOffset

A row which is located by an offset from the last row.

Row

A row which is located by an offset from the first row.

Rows

Row denotes a set of cells on given rows on a Table.

Segment

Segment represents a sub table of Table.

Traits

Object

Object helps to locate a nessesary part of a Table.