Trait tabled::object::Object

pub trait Object: Sized {
    type Iter: Iterator<Item = Entity>;

    // Required method
    fn cells<R>(&self, table: &Table<R>) -> Self::Iter
       where R: Records;

    // Provided methods
    fn and<O>(self, rhs: O) -> UnionCombination<Self, O>
       where O: Object { ... }
    fn not<O>(self, rhs: O) -> DiffCombination<Self, O>
       where O: Object { ... }
    fn intersect<O>(self, rhs: O) -> IntersectionCombination<Self, O>
       where O: Object { ... }
    fn inverse(self) -> InversionCombination<Self> { ... }
}

Object helps to locate a necessary part of a Table.

Required Associated Types

type Iter: Iterator<Item = Entity>

An Iterator which returns a list of cells.

Required Methods

fn cells<R>(&self, table: &Table<R>) -> Self::Iterwhere R: Records,

Cells returns a set of coordinates of cells

Provided Methods

fn and<O>(self, rhs: O) -> UnionCombination<Self, O>where O: Object,

Combines cells. It doesn’t repeat cells.

fn not<O>(self, rhs: O) -> DiffCombination<Self, O>where O: Object,

Excludes rhs cells from this cells.

fn intersect<O>(self, rhs: O) -> IntersectionCombination<Self, O>where O: Object,

Returns cells which are present in both Objects only.

fn inverse(self) -> InversionCombination<Self>

Returns cells which are not present in target Object.