tabled::settings::disable

Struct Remove

pub struct Remove<L, Target> { /* private fields */ }

Remove removes particular rows/columns from a Table.

It tries to keeps track of style changes which may occur. But it’s not guaranteed will be the way you would expect it to be.

Generally you should avoid use of Remove because it’s a slow function and modifies the underlying records. Providing correct data right away is better.

Example

use tabled::{Table, settings::{Remove, object::Rows}};

let data = vec!["Hello", "World", "!!!"];

let table = Table::new(data).with(Remove::row(Rows::new(1..2))).to_string();

assert_eq!(
    table,
    "+-------+\n\
     | &str  |\n\
     +-------+\n\
     | World |\n\
     +-------+\n\
     | !!!   |\n\
     +-------+"
);

Implementations

impl<L> Remove<L, TargetColumn>

pub fn column(locator: L) -> Self

Remove columns.

Available locators are:

Columns

Column

FirstColumn

LastColumn

ByColumnName

use tabled::{builder::Builder, settings::{Remove, location::ByColumnName, object::Columns}};

let mut builder = Builder::default();
builder.push_record(["col1", "col2", "col3"]);
builder.push_record(["Hello", "World", "1"]);

let table = builder.build()
    .with(Remove::column(ByColumnName::new("col3")))
    .to_string();

assert_eq!(
    table,
    "+-------+-------+\n\
     | col1  | col2  |\n\
     +-------+-------+\n\
     | Hello | World |\n\
     +-------+-------+"
);

impl<L> Remove<L, TargetRow>

pub fn row(locator: L) -> Self

Remove rows.

Available locators are:

Rows

Row

FirstRow

LastRow

use tabled::{settings::{Remove, object::Rows}, builder::Builder};

let mut builder = Builder::default();
builder.push_record(["col1", "col2", "col3"]);
builder.push_record(["Hello", "World", "1"]);

let table = builder.build()
    .with(Remove::row(Rows::first()))
    .to_string();

assert_eq!(
    table,
    "+-------+-------+---+\n\
     | Hello | World | 1 |\n\
     +-------+-------+---+"
);