tabled::settings

Module formatting

This module contains settings for render strategy of papergrid.

TrimStrategy and AlignmentStrategy allows to set Alignment settings.

TabSize sets a default tab size.

Charset responsible for special char treatment.

Justification responsible for justification space of content.

Structs

Charset

A structure to handle special chars.

CleanCharset

CleanCharset removes all \t and \r occurrences.

Justification

Set a justification character and a color.

TabSize

Set a tab size.

Enums

AlignmentStrategy

AlignmentStrategy is a responsible for a flow how we apply an alignment. It mostly matters for multiline strings.

TrimStrategy

TrimStrategy determines if it’s allowed to use empty space while doing Alignment.