tabled::settings::highlight

Struct Highlight

pub struct Highlight<O> { /* private fields */ }

Highlight modifies a table style by changing a border of a target Table segment.

It basically highlights outer border of a given segment.

Example

use tabled::{
    Table,
    settings::{
        Highlight, Border, Style,
        object::{Segment, Object}
    }
};

let data = [
    ("ELF", "Extensible Linking Format", true),
    ("DWARF", "", true),
    ("PE", "Portable Executable", false),
];

let table = Table::new(data.iter().enumerate())
               .with(Style::markdown())
               .with(Highlight::outline(Segment::all().not((0,0).and((1, 0)).and((0, 1)).and((0, 3))), '*'))
               .to_string();

assert_eq!(
    table,
    concat!(
        "                *****************************        \n",
        "| usize | &str  * &str                      * bool  |\n",
        "|-------*********---------------------------*********\n",
        "| 0     * ELF   | Extensible Linking Format | true  *\n",
        "*********                                           *\n",
        "* 1     | DWARF |                           | true  *\n",
        "*                                                   *\n",
        "* 2     | PE    | Portable Executable       | false *\n",
        "*****************************************************",
    ),
);

Implementations

impl<O> Highlight<O>

pub const fn new(target: O) -> Self

Build a new instance of Highlight with ‘*’ char being used as changer.

BE AWARE: if target exceeds boundaries it may panic.

pub const fn outline(target: O, c: char) -> Self

Build a new instance of Highlight, highlighting by a character.

BE AWARE: if target exceeds boundaries it may panic.

pub fn colored(target: O, color: Color) -> Self

Build a new instance of Highlight, highlighting by a color and a given character for a border.

BE AWARE: if target exceeds boundaries it may panic.

pub fn border<T, B, L, R>(self, border: ConstBorder<T, B, L, R>) -> Self

Set a border for a Highlight.

pub fn color(self, border: BorderColor) -> Self

Set a border color for a Highlight.