Module tabled::settings::measurement

The module contains Measurement trait and its implementations to be used in Height and Width.;

Structs

Max

Max width value.

Min

Min width value.

Percent

Percent from a total table width.

Traits

Measurement

A width value which can be obtained on behalf of Table.