tabled::settings::object

Trait Object

pub trait Object<R> {
    type Iter: Iterator<Item = Entity>;

    // Required method
    fn cells(&self, records: &R) -> Self::Iter;

    // Provided methods
    fn and<O>(self, rhs: O) -> UnionCombination<Self, O, R>
       where Self: Sized { ... }
    fn not<O>(self, rhs: O) -> DiffCombination<Self, O, R>
       where Self: Sized { ... }
    fn intersect<O>(self, rhs: O) -> IntersectionCombination<Self, O, R>
       where Self: Sized { ... }
    fn inverse(self) -> InversionCombination<Self, R>
       where Self: Sized { ... }
}

Object helps to locate a necessary part of a Table.

Required Associated Types

type Iter: Iterator<Item = Entity>

An Iterator which returns a list of cells.

Required Methods

fn cells(&self, records: &R) -> Self::Iter

Cells returns a set of coordinates of cells.

Provided Methods

fn and<O>(self, rhs: O) -> UnionCombination<Self, O, R>
where Self: Sized,

Combines cells. It doesn’t repeat cells.

fn not<O>(self, rhs: O) -> DiffCombination<Self, O, R>
where Self: Sized,

Excludes rhs cells from this cells.

fn intersect<O>(self, rhs: O) -> IntersectionCombination<Self, O, R>
where Self: Sized,

Returns cells which are present in both Objects only.

fn inverse(self) -> InversionCombination<Self, R>
where Self: Sized,

Returns cells which are not present in target Object.