tabled::settings::peaker

Trait Peaker

pub trait Peaker {
    // Required method
    fn peak(&mut self, mins: &[usize], values: &[usize]) -> Option<usize>;
}

A strategy of width function. It determines the order how a function is applied.

For example which column we shall peak to truncate when doing width alogorithms.

Required Methods

fn peak(&mut self, mins: &[usize], values: &[usize]) -> Option<usize>

This function returns an index which will be changed. Or None if no changes are necessary.

When None returned the alogorithm must be stopped.