tabled::settings::span

Struct Span

pub struct Span;

Span represent a horizontal/column span setting for any cell on a Table.

use tabled::{settings::{Span, Modify}, Table};
use tabled::assert::assert_table;

let data = [[1, 2, 3], [4, 5, 6]];

let mut table = Table::new(data);
table.modify((0, 0), Span::row(2));
table.modify((0, 1), Span::column(2));
table.modify((2, 0), Span::column(1000));

assert_table!(
    table,
    "+---+---+---+"
    "| 0 | 1     |"
    "+   +---+---+"
    "|   | 2 | 3 |"
    "+---+---+---+"
    "| 4         |"
    "+---+---+---+"
);

Implementations

impl Span

pub fn column(size: isize) -> ColumnSpan

New constructs a horizontal/column Span.

Value can be: * == 0 - which means spread the cell on the whole line * == 1 - which is a default span so can be used for removal of spans * > 1 - which means to spread a cell by given number of columns right * < 0 - which means to spread a cell by given number of columns left

Example

use tabled::{settings::{Span, Modify}, Table};

let data = [[1, 2, 3], [4, 5, 6]];

let table = Table::new(data)
    .modify((0, 0), Span::column(100))
    .modify((1, 1), Span::column(2))
    .modify((2, 1), Span::column(-1))
    .to_string();

assert_eq!(
    table,
    concat!(
        "+---++---+\n",
        "| 0      |\n",
        "+---++---+\n",
        "| 1 | 2  |\n",
        "+---++---+\n",
        "| 5  | 6 |\n",
        "+---++---+",
    )
)

pub fn row(size: isize) -> RowSpan

New constructs a vertical/row Span.

Value can be: * == 0 - which means spread the cell on the whole line * == 1 - which is a default span so can be used for removal of spans * > 1 - which means to spread a cell by given number of rows bottom * < 0 - which means to spread a cell by given number of rows top

Example

use tabled::{settings::{Span, Modify}, Table};

let data = [[1, 2, 3], [4, 5, 6]];

let table = Table::new(data)
    .modify((0, 0), Span::row(100))
    .modify((1, 1), Span::row(2))
    .modify((2, 2), Span::row(-1))
    .to_string();

assert_eq!(
    table,
    concat!(
        "+---+---+---+\n",
        "| 0 | 1 | 2 |\n",
        "+   +---+---+\n",
        "+   + 2 + 6 +\n",
        "+---+---+---+",
    )
)