Struct tabled::settings::split::Split

pub struct Split { /* private fields */ }

Returns a new Table formatted with several optional parameters.

The required index parameter determines how many columns/rows a table will be redistributed into.

index

direction

behavior

display

Example

use std::iter::FromIterator;
use tabled::{
    settings::split::Split,
    Table,
};

let mut table = Table::from_iter(['a'..='z']);
let table = table.with(Split::column(4)).to_string();

assert_eq!(table, "+---+---+---+---+\n\
                   | a | b | c | d |\n\
                   +---+---+---+---+\n\
                   | e | f | g | h |\n\
                   +---+---+---+---+\n\
                   | i | j | k | l |\n\
                   +---+---+---+---+\n\
                   | m | n | o | p |\n\
                   +---+---+---+---+\n\
                   | q | r | s | t |\n\
                   +---+---+---+---+\n\
                   | u | v | w | x |\n\
                   +---+---+---+---+\n\
                   | y | z |   |   |\n\
                   +---+---+---+---+")

Implementations

impl Split

pub fn column(index: usize) -> Self

Returns a new Table split on the column at the provided index.

The column found at that index becomes the new right-most column in the returned table. Columns found beyond the index are redistributed into the table based on other defined parameters.

Split::column(4);

pub fn row(index: usize) -> Self

Returns a new Table split on the row at the provided index.

The row found at that index becomes the new bottom row in the returned table. Rows found beyond the index are redistributed into the table based on other defined parameters.

Split::row(4);

pub fn concat(self) -> Self

Returns a split Table with the redistributed cells pushed to the back of the new shape.

                                                +---+---+
                                                | a | b |
                                                +---+---+
+---+---+---+---+---+                           | f | g |
| a | b | c | d | e | Split::column(2).concat() +---+---+
+---+---+---+---+---+           =>              | c | d |
| f | g | h | i | j |                           +---+---+
+---+---+---+---+---+                           | h | i |
                                                +---+---+
                                                | e |   |
                                                +---+---+
                                                | j |   |
                                                +---+---+

pub fn zip(self) -> Self

Returns a split Table with the redistributed cells inserted behind the first correlating column/row one after another.

                                             +---+---+
                                             | a | b |
                                             +---+---+
+---+---+---+---+---+                        | c | d |
| a | b | c | d | e | Split::column(2).zip() +---+---+
+---+---+---+---+---+           =>           | e |   |
| f | g | h | i | j |                        +---+---+
+---+---+---+---+---+                        | f | g |
                                             +---+---+
                                             | h | i |
                                             +---+---+
                                             | j |   |
                                             +---+---+

pub fn clean(self) -> Self

Returns a split Table with the empty columns/rows filtered out.

                                                
                                                
                                               +---+---+---+
+---+---+---+---+---+                          | a | b | c |
| a | b | c | d | e | Split::column(3).clean() +---+---+---+
+---+---+---+---+---+           =>             | d | e |   |
| f | g | h |   |   |                          +---+---+---+
+---+---+---+---+---+                          | f | g | h |
              ^   ^                            +---+---+---+
              these cells are filtered
              from the resulting table

Notes

This is apart of the default configuration for Split.

See retain for an alternative display option.

pub fn retain(self) -> Self

Returns a split Table with all cells retained.

                                                +---+---+---+
                                                | a | b | c |
                                                +---+---+---+
+---+---+---+---+---+                           | d | e |   |
| a | b | c | d | e | Split::column(3).retain() +---+---+---+
+---+---+---+---+---+           =>              | f | g | h |
| f | g | h |   |   |                           +---+---+---+
+---+---+---+---+---+             |-----------> |   |   |   |
              ^   ^               |             +---+---+---+
              |___|_____cells are kept!

Notes

See clean for an alternative display option.