Struct tabled::settings::Alignment

pub struct Alignment { /* private fields */ }

Alignment represent a horizontal and vertical alignment setting for any cell on a Table.

An alignment strategy can be set by AlignmentStrategy.

Example

use tabled::{
    Table,
    settings::{
        formatting::AlignmentStrategy,
        object::Segment, Alignment, Modify, Style,
    }
};

let data = [
    ["1", "2", "3"],
    ["Some\nMulti\nLine\nText", "and a line", "here"],
    ["4", "5", "6"],
];

let mut table = Table::new(&data);
table
    .with(Style::modern())
    .with(
        Modify::new(Segment::all())
            .with(Alignment::right())
            .with(Alignment::center())
            .with(AlignmentStrategy::PerCell)
    );

assert_eq!(
    table.to_string(),
    concat!(
        "┌───────┬────────────┬──────┐\n",
        "│   0   │     1      │  2   │\n",
        "├───────┼────────────┼──────┤\n",
        "│   1   │     2      │  3   │\n",
        "├───────┼────────────┼──────┤\n",
        "│ Some  │ and a line │ here │\n",
        "│ Multi │            │      │\n",
        "│ Line  │            │      │\n",
        "│ Text  │            │      │\n",
        "├───────┼────────────┼──────┤\n",
        "│   4   │     5      │  6   │\n",
        "└───────┴────────────┴──────┘",
    ),
)

Implementations

impl Alignment

pub const fn left() -> Self

Left constructs a horizontal alignment to AlignmentHorizontal::Left

pub const fn right() -> Self

Right constructs a horizontal alignment to AlignmentHorizontal::Right

Notice

When you use MinWidth the alignment might not work as you expected. You could try to apply TrimStrategy which may help.

pub const fn center() -> Self

Center constructs a horizontal alignment to AlignmentHorizontal::Center

Notice

When you use MinWidth the alignment might not work as you expected. You could try to apply TrimStrategy which may help.

pub const fn top() -> Self

Top constructs a vertical alignment to AlignmentVertical::Top

pub const fn bottom() -> Self

Bottom constructs a vertical alignment to AlignmentVertical::Bottom

pub const fn center_vertical() -> Self

Center_vertical constructs a vertical alignment to AlignmentVertical::Center