tabled::settings

Struct Color

pub struct Color { /* private fields */ }

Color represents a color which can be set to things like Border, Padding and Margin.

Example

use tabled::{settings::Color, Table};

let data = [
    (0u8, "Hello"),
    (1u8, "World"),
];

let table = Table::new(data)
    .with(Color::BG_BLUE)
    .to_string();

println!("{}", table);

Implementations

impl Color

pub const FG_BLACK: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BLUE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_BLACK: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_BLUE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_CYAN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_GREEN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_MAGENTA: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_RED: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_WHITE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_BRIGHT_YELLOW: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_CYAN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_GREEN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_MAGENTA: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_RED: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_WHITE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const FG_YELLOW: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BLACK: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BLUE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_BLACK: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_BLUE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_CYAN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_GREEN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_MAGENTA: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_RED: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_WHITE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_BRIGHT_YELLOW: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_CYAN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_GREEN: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_MAGENTA: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_RED: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_WHITE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BG_YELLOW: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const BOLD: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

pub const UNDERLINE: Self

A color representation.

Notice that the colors are constants so you can’t combine them.

impl Color

pub fn new<P, S>(prefix: P, suffix: S) -> Self
where P: Into<String>, S: Into<String>,

Creates a new Color instance, with ANSI prefix and ANSI suffix. You can use [TryFrom] to construct it from [String`].

pub const fn empty() -> Self

Creates a new empty Color`.

pub fn get_prefix(&self) -> &str

Return a prefix.

pub fn get_suffix(&self) -> &str

Return a suffix.

pub fn as_ansi_str(&self) -> Option<StaticColor<'static>>

Tries to get a static value of the color.

pub fn parse<S>(text: S) -> Self
where S: AsRef<str>,

Parses the string,

Panics

PANICS if the input string incorrectly built. Use std::convert::TryFrom instead if you are not sure about the input.

pub fn rgb_fg(r: u8, g: u8, b: u8) -> Self

Create a 24 bit foreground color with RGB

pub fn rgb_bg(r: u8, g: u8, b: u8) -> Self

Create a 24 bit background color with RGB.

The terminal need to support the escape sequence

pub fn colorize<S>(&self, text: S) -> String
where S: AsRef<str>,

Colorize a string.