Module tabled::settings::style

This module contains a list of primitives which can be applied to change Table style.

Style

It is responsible for a table border style. An individual cell border can be set by Border.

Example

use tabled::{Table, settings::Style};

let data = vec!["Hello", "2022"];
let mut table = Table::new(&data);
table.with(Style::psql());

assert_eq!(
    table.to_string(),
    concat!(
        " &str  \n",
        "-------\n",
        " Hello \n",
        " 2022  ",
    )
)

BorderText

It’s used to override a border with a custom text.

Example

use tabled::{Table, settings::style::{BorderText, Style}};

let data = vec!["Hello", "2022"];
let table = Table::new(&data)
    .with(Style::psql())
    .with(BorderText::new("Santa").horizontal(1))
    .to_string();

assert_eq!(
    table,
    concat!(
        " &str  \n",
        "Santa--\n",
        " Hello \n",
        " 2022  ",
    )
)

Border

Border can be used to modify cell’s borders.

It’s possible to set a collored border when color feature is on.

Example

use tabled::{Table, settings::{Modify, Style}};

let data = vec!["Hello", "2022"];
let table = Table::new(&data)
    .with(Style::psql())
    .with(Modify::new((0, 0)).with(Style::modern().get_frame()))
    .to_string();

assert_eq!(
    table,
    concat!(
        "┌───────┐\n",
        "│ &str  │\n",
        "└───────┘\n",
        "  Hello  \n",
        "  2022   ",
    )
)

RawStyle

A different representation of Style. With no checks in place.

It also contains a list of types to support colors.

Structs

Border

Border represents a border of a Cell.

BorderChar

BorderChar sets a char to a specific location on a horizontal line.

BorderColor

BorderColored represents a colored border of a Cell.

BorderSpanCorrection

A correctness function of style for Table which has Spans.

BorderText

BorderText writes a custom text on a border.

HorizontalLine

A horizontal split line which can be used to set a border.

HorizontalLineIter

An iterator which limits Line influence on iterations over lines for in Style.

Line

The structure represent a vertical or horizontal line.

On

A marker struct which is used in Style.

RawStyle

A raw style data, which can be produced safely from Style.

Style

Style is represents a theme of a Table.

VerticalLine

A horizontal split line which can be used to set a border.

VerticalLineIter

An iterator which limits Line influence on iterations over lines for in Style.

Enums

Offset

The structure represents an offset in a text.