Struct tabled::settings::style::BorderColor

pub struct BorderColor(/* private fields */);

BorderColored represents a colored border of a Cell.

let table = Table::new(&data)
    .with(Style::ascii())
    .with(Modify::new(Rows::single(0)).with(BorderColor::default().top(Color::FG_RED)));

Implementations

impl BorderColor

pub fn full( top: Color, bottom: Color, left: Color, right: Color, top_left: Color, top_right: Color, bottom_left: Color, bottom_right: Color ) -> Self

This function constructs a cell borders with all sides set.

pub fn filled(c: Color) -> Self

This function constructs a cell borders with all sides’s char set to a given character. It behaves like Border::full with the same character set to each side.

pub fn top(self, c: Color) -> Self

Set a top border character.

pub fn bottom(self, c: Color) -> Self

Set a bottom border character.

pub fn left(self, c: Color) -> Self

Set a left border character.

pub fn right(self, c: Color) -> Self

Set a right border character.

pub fn corner_top_left(self, c: Color) -> Self

Set a top left intersection character.

pub fn corner_top_right(self, c: Color) -> Self

Set a top right intersection character.

pub fn corner_bottom_left(self, c: Color) -> Self

Set a bottom left intersection character.

pub fn corner_bottom_right(self, c: Color) -> Self

Set a bottom right intersection character.