Struct tabled::settings::style::BorderText

pub struct BorderText<L> { /* private fields */ }

BorderText writes a custom text on a border.

Example

use tabled::{Table, settings::style::BorderText, settings::object::Rows};

let mut table = Table::new(["Hello World"]);
table
    .with(BorderText::new("+-.table").horizontal(Rows::first()));

assert_eq!(
    table.to_string(),
    "+-.table------+\n\
     | &str        |\n\
     +-------------+\n\
     | Hello World |\n\
     +-------------+"
);

Implementations

impl BorderText<()>

pub fn new<S: Into<String>>(text: S) -> Self

Creates a BorderText instance.

Lines are numbered from 0 to the count_rows included (line >= 0 && line <= count_rows).

impl<Line> BorderText<Line>

pub fn horizontal<L>(self, line: L) -> BorderText<L>

Set a line on which we will set the text.

pub fn offset(self, offset: Offset) -> Self

Set an offset from which the text will be started.

pub fn color(self, color: Color) -> Self

Set a color of the text.