Struct tabled::settings::style::RawStyle

pub struct RawStyle { /* private fields */ }

A raw style data, which can be produced safely from Style.

It can be useful in order to not have a generics and be able to use it as a variable more conveniently.

Implementations

impl RawStyle

pub fn set_top(&mut self, s: Option<char>) -> &mut Self

Set a top border character.

pub fn set_color_top(&mut self, color: Color) -> &mut Self

Set a top border color.

pub fn set_bottom(&mut self, s: Option<char>) -> &mut Self

Set a bottom border character.

pub fn set_color_bottom(&mut self, color: Color) -> &mut Self

Set a bottom border color.

pub fn set_left(&mut self, s: Option<char>) -> &mut Self

Set a left border character.

pub fn set_color_left(&mut self, color: Color) -> &mut Self

Set a left border color.

pub fn set_right(&mut self, s: Option<char>) -> &mut Self

Set a right border character.

pub fn set_color_right(&mut self, color: Color) -> &mut Self

Set a right border color.

pub fn set_intersection_top(&mut self, s: Option<char>) -> &mut Self

Set a top intersection character.

pub fn set_color_intersection_top(&mut self, color: Color) -> &mut Self

Set a top intersection color.

pub fn set_intersection_bottom(&mut self, s: Option<char>) -> &mut Self

Set a bottom intersection character.

pub fn set_color_intersection_bottom(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_intersection_left(&mut self, s: Option<char>) -> &mut Self

Set a left split character.

pub fn set_color_intersection_left(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_intersection_right(&mut self, s: Option<char>) -> &mut Self

Set a right split character.

pub fn set_color_intersection_right(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_intersection(&mut self, s: Option<char>) -> &mut Self

Set an internal character.

pub fn set_color_intersection(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_vertical(&mut self, s: Option<char>) -> &mut Self

Set a vertical character.

pub fn set_color_vertical(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_horizontal(&mut self, s: Option<char>) -> &mut Self

Set a horizontal character.

pub fn set_color_horizontal(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_corner_top_left(&mut self, s: Option<char>) -> &mut Self

Set a character for a top left corner.

pub fn set_color_corner_top_left(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_corner_top_right(&mut self, s: Option<char>) -> &mut Self

Set a character for a top right corner.

pub fn set_color_corner_top_right(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_corner_bottom_left(&mut self, s: Option<char>) -> &mut Self

Set a character for a bottom left corner.

pub fn set_color_corner_bottom_left(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_corner_bottom_right(&mut self, s: Option<char>) -> &mut Self

Set a character for a bottom right corner.

pub fn set_color_corner_bottom_right(&mut self, color: Color) -> &mut Self

Set a bottom intersection color.

pub fn set_horizontals(&mut self, lines: HashMap<usize, Line>) -> &mut Self

Set horizontal border lines.

Example

use std::collections::HashMap;
use tabled::{Table, settings::style::{Style, Line, RawStyle}};

let mut style = RawStyle::from(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, Style::extended().get_horizontal());
style.set_horizontals(lines);

let table = Table::new((0..3).map(|i| ("Hello", i)))
   .with(style)
   .to_string();

assert_eq!(
    table,
    concat!(
        " ======= ===== \n",
        "  &str    i32  \n",
        "╠═══════╬═════╣\n",
        "  Hello   0    \n",
        "  Hello   1    \n",
        "  Hello   2    \n",
        " ======= ===== ",
    ),
)

pub fn insert_horizontal(&mut self, row: usize, line: Line) -> &mut Self

Insert a horizontal line to a specific row location.

pub fn get_horizontal(&self, row: usize) -> Option<Line>

Insert a horizontal line to a specific row location.

pub fn set_verticals(&mut self, lines: HashMap<usize, Line>) -> &mut Self

Set vertical border lines.

Example

use std::collections::HashMap;
use tabled::{Table, settings::style::{Style, Line, RawStyle}};

let mut style = RawStyle::from(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, Style::extended().get_horizontal());
style.set_verticals(lines);

let table = Table::new((0..3).map(|i| ("Hello", i)))
   .with(style)
   .to_string();

assert_eq!(
    table,
    concat!(
        "=======╠=====\n",
        " &str  ═ i32 \n",
        "======= =====\n",
        " Hello ═ 0   \n",
        " Hello ═ 1   \n",
        " Hello ═ 2   \n",
        "=======╣=====",
    ),
)

pub fn insert_vertical(&mut self, column: usize, line: Line) -> &mut Self

Insert a vertical line into specific column location.

pub fn get_left(&self) -> Option<char>

Get a left char.

pub fn get_left_intersection(&self) -> Option<char>

Get a left intersection char.

pub fn get_right(&self) -> Option<char>

Get a right char.

pub fn get_right_intersection(&self) -> Option<char>

Get a right intersection char.

pub fn get_top(&self) -> Option<char>

Get a top char.

pub fn get_top_left(&self) -> Option<char>

Get a top left char.

pub fn get_top_right(&self) -> Option<char>

Get a top right char.

pub fn get_top_intersection(&self) -> Option<char>

Get a top intersection char.

pub fn get_bottom(&self) -> Option<char>

Get a bottom intersection char.

pub fn get_bottom_left(&self) -> Option<char>

Get a bottom intersection char.

pub fn get_bottom_right(&self) -> Option<char>

Get a bottom intersection char.

pub fn get_bottom_intersection(&self) -> Option<char>

Get a bottom intersection char.

pub fn get_frame(&self) -> Border

Returns an outer border of the style.

pub fn get_borders(&self) -> Borders<char>

Returns an general borders configuration of the style.