Struct tabled::settings::style::Style

pub struct Style<T, B, L, R, H, V, HLines = [HorizontalLine; N], VLines = [VerticalLine; N]> { /* private fields */ }

Style is represents a theme of a Table.

corner top left         top intersection                    corner top right
               .             |                             .
                .            V                            .
                 ╭───┬───┬───┬───┬───┬───┬────┬────┬────╮
                 │ i │ 0 │ 1 │ 2 │ 3 │ 4 │ 5  │ 6  │ 7  │
                 ├───┼───┼───┼───┼───┼───┼────┼────┼────┤ <- this horizontal line is custom 'horizontals'
                 │ 0 │ 0 │ 0 │ 0 │ 0 │ 0 │ 0  │ 0  │ 0  │    other lines horizontal lines are not set they called 'horizontal'
                 │ 1 │ 0 │ 1 │ 2 │ 3 │ 4 │ 5  │ 6  │ 7  │
                 │ 2 │ 0 │ 2 │ 4 │ 6 │ 8 │ 10 │ 12 │ 14 │
                 ╰───┴───┴───┴───┴───┴───┴────┴────┴────╯
               .         ^                    ^           .
              .          |                    |            .
corner bottom left       |         bottom intersection     corner bottom right
                         |
                         |
            all this vertical lines are called 'vertical'

                    ┌───┬───┬───┬───┬───┐
                    │ 0 │ 1 │ 2 │ 3 │ 4 │
intersection left ->├───X───X───X───X───┤ <- all this horizontal lines are called 'horizontal'
                    │ 1 │ 2 │ 3 │ 4 │ 5 │
                    ├───X───X───X───X───┤ <- intersection right
                    │ 2 │ 3 │ 4 │ 5 │ 6 │
                    └───┴───┴───┴───┴───┘

All 'X' positions are called 'intersection'.
It's a place where 'vertical' and 'horizontal' lines intersect.

It tries to limit an controlling a valid state of it. For example, it won’t allow to call method Style::corner_top_left unless Style::left and Style::top is set.

You can turn Style into RawStyle to have more control using Into implementation.

Example

use tabled::{Table, settings::Style};

let style = Style::ascii()
                .bottom('*')
                .intersection(' ');

let data = vec!["Hello", "2021"];
let table = Table::new(&data).with(style).to_string();

println!("{}", table);

Implementations

impl Style<(), (), (), (), (), (), (), ()>

pub const fn empty() -> Style<(), (), (), (), (), ()>

This style is a style with no styling options on,

     id  destribution            link
     0      Fedora      https://getfedora.org/
     2     OpenSUSE    https://www.opensuse.org/
     3   Endeavouros   https://endeavouros.com/

Note: The cells in the example have 1-left and 1-right indent.

This style can be used as a base style to build a custom one.

let style = Style::empty()
    .top('*')
    .bottom('*')
    .vertical('#')
    .intersection_top('*');

pub const fn blank() -> Style<(), (), (), (), (), On>

This style is analog of empty but with a vertical space(’ ’) line.

     id   destribution             link
     0       Fedora       https://getfedora.org/
     2      OpenSUSE     https://www.opensuse.org/
     3    Endeavouros    https://endeavouros.com/

pub const fn ascii() -> Style<On, On, On, On, On, On>

This is a style which relays only on ASCII charset.

It has horizontal and vertical lines.

    +----+--------------+---------------------------+
    | id | destribution |           link            |
    +----+--------------+---------------------------+
    | 0  |    Fedora    |  https://getfedora.org/   |
    +----+--------------+---------------------------+
    | 2  |   OpenSUSE   | https://www.opensuse.org/ |
    +----+--------------+---------------------------+
    | 3  | Endeavouros  | https://endeavouros.com/  |
    +----+--------------+---------------------------+

pub const fn psql() -> Style<(), (), (), (), (), On, [HorizontalLine; N]>

psql style looks like a table style PostgreSQL uses.

It has only 1 horizontal line which splits header. And no left and right vertical lines.

     id | destribution |           link
    ----+--------------+---------------------------
     0  |    Fedora    |  https://getfedora.org/
     2  |   OpenSUSE   | https://www.opensuse.org/
     3  | Endeavouros  | https://endeavouros.com/

pub const fn markdown() -> Style<(), (), On, On, (), On, [HorizontalLine; N]>

markdown style mimics a Markdown table style.

    | id | destribution |           link            |
    |----|--------------|---------------------------|
    | 0  |    Fedora    |  https://getfedora.org/   |
    | 2  |   OpenSUSE   | https://www.opensuse.org/ |
    | 3  | Endeavouros  | https://endeavouros.com/  |

pub const fn modern() -> Style<On, On, On, On, On, On>

This style is analog of Style::ascii which uses UTF-8 charset.

It has vertical and horizontal split lines.

    ┌────┬──────────────┬───────────────────────────┐
    │ id │ destribution │           link            │
    ├────┼──────────────┼───────────────────────────┤
    │ 0  │    Fedora    │  https://getfedora.org/   │
    ├────┼──────────────┼───────────────────────────┤
    │ 2  │   OpenSUSE   │ https://www.opensuse.org/ │
    ├────┼──────────────┼───────────────────────────┤
    │ 3  │ Endeavouros  │ https://endeavouros.com/  │
    └────┴──────────────┴───────────────────────────┘

pub const fn sharp() -> Style<On, On, On, On, (), On, [HorizontalLine; N]>

This style looks like a Style::modern but without horozizontal lines except a header.

Beware: It uses UTF-8 characters.

    ┌────┬──────────────┬───────────────────────────┐
    │ id │ destribution │           link            │
    ├────┼──────────────┼───────────────────────────┤
    │ 0  │    Fedora    │  https://getfedora.org/   │
    │ 2  │   OpenSUSE   │ https://www.opensuse.org/ │
    │ 3  │ Endeavouros  │ https://endeavouros.com/  │
    └────┴──────────────┴───────────────────────────┘

pub const fn rounded() -> Style<On, On, On, On, (), On, [HorizontalLine; N]>

This style looks like a Style::sharp but with rounded corners.

Beware: It uses UTF-8 characters.

    ╭────┬──────────────┬───────────────────────────╮
    │ id │ destribution │           link            │
    ├────┼──────────────┼───────────────────────────┤
    │ 0  │    Fedora    │  https://getfedora.org/   │
    │ 2  │   OpenSUSE   │ https://www.opensuse.org/ │
    │ 3  │ Endeavouros  │ https://endeavouros.com/  │
    ╰────┴──────────────┴───────────────────────────╯

pub const fn extended() -> Style<On, On, On, On, On, On>

This style uses a chars which resembles ‘2 lines’.

Beware: It uses UTF8 characters.

    ╔════╦══════════════╦═══════════════════════════╗
    ║ id ║ destribution ║           link            ║
    ╠════╬══════════════╬═══════════════════════════╣
    ║ 0  ║    Fedora    ║  https://getfedora.org/   ║
    ╠════╬══════════════╬═══════════════════════════╣
    ║ 2  ║   OpenSUSE   ║ https://www.opensuse.org/ ║
    ╠════╬══════════════╬═══════════════════════════╣
    ║ 3  ║ Endeavouros  ║ https://endeavouros.com/  ║
    ╚════╩══════════════╩═══════════════════════════╝

pub const fn dots() -> Style<On, On, On, On, On, On>

This is a style uses only ‘.’ and ‘:’ chars. It has a vertical and horizontal split lines.

    .................................................
    : id : destribution :           link            :
    :....:..............:...........................:
    : 0  :    Fedora    :  https://getfedora.org/   :
    :....:..............:...........................:
    : 2  :   OpenSUSE   : https://www.opensuse.org/ :
    :....:..............:...........................:
    : 3  : Endeavouros  : https://endeavouros.com/  :
    :....:..............:...........................:

pub const fn re_structured_text( ) -> Style<On, On, (), (), (), On, [HorizontalLine; N]>

This style is one of table views in ReStructuredText.

    ==== ============== ===========================
     id   destribution             link            
    ==== ============== ===========================
     0       Fedora       https://getfedora.org/   
     2      OpenSUSE     https://www.opensuse.org/
     3    Endeavouros    https://endeavouros.com/  
    ==== ============== ===========================

pub const fn ascii_rounded() -> Style<On, On, On, On, (), On>

This is a theme analog of Style::rounded, but in using ascii charset and with no horizontal lines.

    .-----------------------------------------------.
    | id | destribution |           link            |
    | 0  |    Fedora    |  https://getfedora.org/   |
    | 2  |   OpenSUSE   | https://www.opensuse.org/ |
    | 3  | Endeavouros  | https://endeavouros.com/  |
    '-----------------------------------------------'

impl<T, B, L, R, H, V, HLines, VLines> Style<T, B, L, R, H, V, HLines, VLines>

pub const fn get_frame(&self) -> Border

Frame function returns a frame as a border.

Example

use tabled::{Table, settings::{Style, Highlight, object::Rows}};

let data = [["10:52:19", "Hello"], ["10:52:20", "World"]];
let table = Table::new(data)
    .with(Highlight::new(Rows::first(), Style::modern().get_frame()))
    .to_string();

assert_eq!(
    table,
    concat!(
        "┌──────────────────┐\n",
        "│ 0        | 1     │\n",
        "└──────────────────┘\n",
        "| 10:52:19 | Hello |\n",
        "+----------+-------+\n",
        "| 10:52:20 | World |\n",
        "+----------+-------+",
    )
);

pub const fn get_horizontal(&self) -> Line

Get a Style’s default horizontal line.

It doesn’t return an overloaded line via Style::horizontals.

Example

use tabled::{settings::style::{Style, HorizontalLine, Line}, Table};

let table = Table::new((0..3).map(|i| ("Hello", "World", i)))
   .with(Style::ascii().remove_horizontal().horizontals([HorizontalLine::new(1, Style::modern().get_horizontal())]))
   .to_string();

assert_eq!(
    table,
    concat!(
        "+-------+-------+-----+\n",
        "| &str  | &str  | i32 |\n",
        "├───────┼───────┼─────┤\n",
        "| Hello | World | 0   |\n",
        "| Hello | World | 1   |\n",
        "| Hello | World | 2   |\n",
        "+-------+-------+-----+",
    )
)

pub const fn get_vertical(&self) -> Line

Get a Style’s default horizontal line.

It doesn’t return an overloaded line via Style::verticals.

Example

use tabled::{settings::style::{Style, VerticalLine, Line}, Table};

let table = Table::new((0..3).map(|i| ("Hello", "World", i)))
   .with(Style::ascii().remove_horizontal().verticals([VerticalLine::new(1, Style::modern().get_vertical())]))
   .to_string();

assert_eq!(
    table,
    concat!(
        "+-------┬-------+-----+\n",
        "| &str  │ &str  | i32 |\n",
        "| Hello │ World | 0   |\n",
        "| Hello │ World | 1   |\n",
        "| Hello │ World | 2   |\n",
        "+-------┴-------+-----+",
    )
)

pub fn top(self, c: char) -> Style<On, B, L, R, H, V, HLines, VLines>where for<'a> &'a mut VLines: IntoIterator<Item = &'a mut VerticalLine>,

Sets a top border.

Any corners and intersections which were set will be overridden.

pub fn bottom(self, c: char) -> Style<T, On, L, R, H, V, HLines, VLines>where for<'a> &'a mut VLines: IntoIterator<Item = &'a mut VerticalLine>,

Sets a bottom border.

Any corners and intersections which were set will be overridden.

pub fn left(self, c: char) -> Style<T, B, On, R, H, V, HLines, VLines>where for<'a> &'a mut HLines: IntoIterator<Item = &'a mut HorizontalLine>,

Sets a left border.

Any corners and intersections which were set will be overridden.

pub fn right(self, c: char) -> Style<T, B, L, On, H, V, HLines, VLines>where for<'a> &'a mut HLines: IntoIterator<Item = &'a mut HorizontalLine>,

Sets a right border.

Any corners and intersections which were set will be overridden.

pub fn horizontal(self, c: char) -> Style<T, B, L, R, On, V, HLines, VLines>where for<'a> &'a mut VLines: IntoIterator<Item = &'a mut VerticalLine>,

Sets a horizontal split line.

Any corners and intersections which were set will be overridden.

pub fn vertical(self, c: char) -> Style<T, B, L, R, H, On, HLines, VLines>where for<'a> &'a mut HLines: IntoIterator<Item = &'a mut HorizontalLine>,

Sets a vertical split line.

Any corners and intersections which were set will be overridden.

pub fn horizontals<NewLines>( self, lines: NewLines ) -> Style<T, B, L, R, H, V, NewLines, VLines>where NewLines: IntoIterator<Item = HorizontalLine> + Clone,

Set border horizontal lines.

Example

use tabled::{settings::style::{Style, HorizontalLine, Line}, Table};

let table = Table::new((0..3).map(|i| ("Hello", i)))
   .with(Style::rounded().horizontals((1..4).map(|i| HorizontalLine::new(i, Line::filled('#')))))
   .to_string();

assert_eq!(
    table,
    concat!(
        "╭───────┬─────╮\n",
        "│ &str  │ i32 │\n",
        "###############\n",
        "│ Hello │ 0   │\n",
        "###############\n",
        "│ Hello │ 1   │\n",
        "###############\n",
        "│ Hello │ 2   │\n",
        "╰───────┴─────╯",
    )
)

pub fn verticals<NewLines>( self, lines: NewLines ) -> Style<T, B, L, R, H, V, HLines, NewLines>where NewLines: IntoIterator<Item = VerticalLine> + Clone,

Set border vertical lines.

Example

use tabled::{Table, settings::style::{Style, VerticalLine, Line}};

let table = Table::new((0..3).map(|i| ("Hello", i)))
   .with(Style::rounded().verticals((0..3).map(|i| VerticalLine::new(i, Line::filled('#')))))
   .to_string();

assert_eq!(
    table,
    concat!(
        "#───────#─────#\n",
        "# &str  # i32 #\n",
        "├───────┼─────┤\n",
        "# Hello # 0   #\n",
        "# Hello # 1   #\n",
        "# Hello # 2   #\n",
        "#───────#─────#",
    )
)

pub fn remove_horizontals( self ) -> Style<T, B, L, R, H, V, [HorizontalLine; N], VLines>

Removes all horizontal lines set by Style::horizontals

pub fn remove_verticals( self ) -> Style<T, B, L, R, H, V, HLines, [VerticalLine; N]>

Removes all verticals lines set by Style::verticals

impl<B, R, H, V, HLines, VLines> Style<On, B, On, R, H, V, HLines, VLines>

pub fn corner_top_left(self, c: char) -> Self

Sets a top left corner.

impl<B, L, H, V, HLines, VLines> Style<On, B, L, On, H, V, HLines, VLines>

pub fn corner_top_right(self, c: char) -> Self

Sets a top right corner.

impl<T, L, H, V, HLines, VLines> Style<T, On, L, On, H, V, HLines, VLines>

pub fn corner_bottom_right(self, c: char) -> Self

Sets a bottom right corner.

impl<T, R, H, V, HLines, VLines> Style<T, On, On, R, H, V, HLines, VLines>

pub fn corner_bottom_left(self, c: char) -> Self

Sets a bottom left corner.

impl<T, B, R, V, HLines, VLines> Style<T, B, On, R, On, V, HLines, VLines>

pub fn intersection_left(self, c: char) -> Self

Sets a left intersection char.

impl<T, B, L, V, HLines, VLines> Style<T, B, L, On, On, V, HLines, VLines>

pub fn intersection_right(self, c: char) -> Self

Sets a right intersection char.

impl<B, L, R, H, HLines, VLines> Style<On, B, L, R, H, On, HLines, VLines>

pub fn intersection_top(self, c: char) -> Self

Sets a top intersection char.

impl<T, L, R, H, HLines, VLines> Style<T, On, L, R, H, On, HLines, VLines>

pub fn intersection_bottom(self, c: char) -> Self

Sets a bottom intersection char.

impl<T, B, L, R, HLines, VLines> Style<T, B, L, R, On, On, HLines, VLines>

pub fn intersection(self, c: char) -> Self

Sets an inner intersection char. A char between horizontal and vertical split lines.

impl<B, L, R, H, V, HLines, VLines> Style<On, B, L, R, H, V, HLines, VLines>

pub fn remove_top( self ) -> Style<(), B, L, R, H, V, HLines, VerticalLineIter<VLines::IntoIter>>where VLines: IntoIterator<Item = VerticalLine> + Clone,

Removes top border.

impl<T, L, R, H, V, HLines, VLines> Style<T, On, L, R, H, V, HLines, VLines>

pub fn remove_bottom( self ) -> Style<T, (), L, R, H, V, HLines, VerticalLineIter<VLines::IntoIter>>where VLines: IntoIterator<Item = VerticalLine> + Clone,

Removes bottom border.

impl<T, B, R, H, V, HLines, VLines> Style<T, B, On, R, H, V, HLines, VLines>

pub fn remove_left( self ) -> Style<T, B, (), R, H, V, HorizontalLineIter<HLines::IntoIter>, VLines>where HLines: IntoIterator<Item = HorizontalLine> + Clone,

Removes left border.

impl<T, B, L, H, V, HLines, VLines> Style<T, B, L, On, H, V, HLines, VLines>

pub fn remove_right( self ) -> Style<T, B, L, (), H, V, HorizontalLineIter<HLines::IntoIter>, VLines>where HLines: IntoIterator<Item = HorizontalLine> + Clone,

Removes right border.

impl<T, B, L, R, V, HLines, VLines> Style<T, B, L, R, On, V, HLines, VLines>

pub fn remove_horizontal( self ) -> Style<T, B, L, R, (), V, HLines, VerticalLineIter<VLines::IntoIter>>where VLines: IntoIterator<Item = VerticalLine> + Clone,

Removes horizontal split lines.

Not including custom split lines.

impl<T, B, L, R, H, HLines, VLines> Style<T, B, L, R, H, On, HLines, VLines>

pub fn remove_vertical( self ) -> Style<T, B, L, R, H, (), HorizontalLineIter<HLines::IntoIter>, VLines>where HLines: IntoIterator<Item = HorizontalLine> + Clone,

Removes vertical split lines.

impl<T, B, L, R, H, V, HLines, VLines> Style<T, B, L, R, H, V, HLines, VLines>

pub const fn get_borders(&self) -> &Borders<char>

Return borders of a table.

pub const fn get_horizontals(&self) -> &HLines

Return custom horizontals which were set.

pub const fn get_verticals(&self) -> &VLines

Return custom verticals which were set.