Struct tabled::settings::themes::Theme

pub struct Theme { /* private fields */ }

A raw style data, which can be produced safely from Style.

It can be useful in order to not have a generics and be able to use it as a variable more conveniently.

Implementations

impl Theme

pub const fn new() -> Self

Creates a new empty style.

It’s an analog of Style::empty

pub const fn from_style<T, B, L, R, H, V, const HS: usize, const VS: usize>( style: Style<T, B, L, R, H, V, HS, VS>, ) -> Self

Build a theme out of a style builder.

pub fn set_frame(&mut self, frame: Border<char>)

Returns an outer border of the style.

pub fn set_frame_colors(&mut self, frame: Border<Color>)

Returns an outer border of the style.

pub fn set_borders(&mut self, borders: Borders<char>)

Set borders structure.

pub fn set_colors(&mut self, borders: Borders<Color>)

Set borders structure.

pub const fn get_borders(&self) -> &Borders<char>

Get borders structure.

pub const fn get_borders_colors(&self) -> &Borders<Color>

Get borders color structure.

pub fn get_borders_mut(&mut self) -> &mut Borders<char>

Get borders structure.

pub fn get_colors_mut(&mut self) -> &mut Borders<Color>

Get borders color structure.

pub fn remove_borders(&mut self)

Remove borders.

pub fn remove_colors(&mut self)

Remove colors.

pub fn remove_horizontal_lines(&mut self)

Remove horizontal lines.

pub fn remove_vertical_lines(&mut self)

Remove vertical lines.

pub const fn get_frame(&self) -> Border<char>

Set an outer border.

pub const fn get_frame_colors(&self) -> Border<&Color>

Set an outer border.

pub fn set_horizontal_lines( &mut self, lines: HashMap<usize, HorizontalLine<char>>, )

Set horizontal border lines.

Example

use std::collections::HashMap;
use tabled::{Table, settings::style::{Style, HorizontalLine}, settings::themes::Theme};

let mut style = Theme::from(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, HorizontalLine::inherit(Style::extended()).into());

style.set_horizontal_lines(lines);

let data = (0..3).map(|i| ("Hello", i));
let table = Table::new(data).with(style).to_string();

assert_eq!(
    table,
    concat!(
        " ======= ===== \n",
        "  &str    i32  \n",
        "╠═══════╬═════╣\n",
        "  Hello   0    \n",
        "  Hello   1    \n",
        "  Hello   2    \n",
        " ======= ===== ",
    ),
)

pub fn set_vertical_lines(&mut self, lines: HashMap<usize, VerticalLine<char>>)

Set vertical border lines.

Example

use std::collections::HashMap;
use tabled::{
    Table,
    settings::style::{Style, HorizontalLine},
    settings::themes::Theme,
};

let mut style = Theme::from_style(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, HorizontalLine::inherit(Style::extended()).into());

style.set_vertical_lines(lines);

let data = (0..3).map(|i| ("Hello", i));
let table = Table::new(data).with(style).to_string();

assert_eq!(
    table,
    concat!(
        "=======╠=====\n",
        " &str  ═ i32 \n",
        "======= =====\n",
        " Hello ═ 0   \n",
        " Hello ═ 1   \n",
        " Hello ═ 2   \n",
        "=======╣=====",
    ),
)

pub fn insert_vertical_line<L>(&mut self, line: usize, vertical: L)
where L: Into<VerticalLine<char>>,

Insert a vertical line into specific column location.

pub fn insert_horizontal_line<L>(&mut self, line: usize, horizontal: L)
where L: Into<HorizontalLine<char>>,

Insert a horizontal line to a specific row location.

pub fn get_vertical_line(&self, column: usize) -> Option<&VerticalLine<char>>

Get a vertical line at the row if any set.

pub fn get_horizontal_line(&self, row: usize) -> Option<&HorizontalLine<char>>

Get a horizontal line at the row if any set.

pub const fn borders_has_left(&self) -> bool

Verifies if borders has left line set on the frame.

pub const fn borders_has_right(&self) -> bool

Verifies if borders has right line set on the frame.

pub const fn borders_has_top(&self) -> bool

Verifies if borders has top line set on the frame.

pub const fn borders_has_bottom(&self) -> bool

Verifies if borders has bottom line set on the frame.

pub const fn borders_has_horizontal(&self) -> bool

Verifies if borders has horizontal lines set.

pub const fn borders_has_vertical(&self) -> bool

Verifies if borders has vertical lines set.

impl Theme

pub fn set_borders_top(&mut self, c: char)

Set a border character top .

pub fn set_borders_bottom(&mut self, c: char)

Set a border character bottom .

pub fn set_borders_left(&mut self, c: char)

Set a border character left .

pub fn set_borders_right(&mut self, c: char)

Set a border character right .

pub fn set_borders_corner_top_left(&mut self, c: char)

Set a border character top left corner .

pub fn set_borders_corner_top_right(&mut self, c: char)

Set a border character top right corner .

pub fn set_borders_corner_bottom_left(&mut self, c: char)

Set a border character bottom left corner .

pub fn set_borders_corner_bottom_right(&mut self, c: char)

Set a border character bottom right corner .

pub fn set_borders_intersection_top(&mut self, c: char)

Set a border character top intersection with a vertical line .

pub fn set_borders_intersection_bottom(&mut self, c: char)

Set a border character bottom intersection with a vertical line .

pub fn set_borders_intersection_left(&mut self, c: char)

Set a border character left intersection with a horizontal line .

pub fn set_borders_intersection_right(&mut self, c: char)

Set a border character right intersection with a horizontal line .

pub fn set_borders_intersection(&mut self, c: char)

Set a border character intersection of horizontal and vertical line .

pub fn set_borders_horizontal(&mut self, c: char)

Set a border character horizontal .

pub fn set_borders_vertical(&mut self, c: char)

Set a border character vertical .

impl Theme

pub const fn get_borders_top(&self) -> Option<char>

Get a border character top .

pub const fn get_borders_bottom(&self) -> Option<char>

Get a border character bottom .

pub const fn get_borders_left(&self) -> Option<char>

Get a border character left .

pub const fn get_borders_right(&self) -> Option<char>

Get a border character right .

pub const fn get_borders_corner_top_left(&self) -> Option<char>

Get a border character top left corner .

pub const fn get_borders_corner_top_right(&self) -> Option<char>

Get a border character top right corner .

pub const fn get_borders_corner_bottom_left(&self) -> Option<char>

Get a border character bottom left corner .

pub const fn get_borders_corner_bottom_right(&self) -> Option<char>

Get a border character bottom right corner .

pub const fn get_borders_intersection_top(&self) -> Option<char>

Get a border character top intersection with a vertical line .

pub const fn get_borders_intersection_bottom(&self) -> Option<char>

Get a border character bottom intersection with a vertical line .

pub const fn get_borders_intersection_left(&self) -> Option<char>

Get a border character left intersection with a horizontal line .

pub const fn get_borders_intersection_right(&self) -> Option<char>

Get a border character right intersection with a horizontal line .

pub const fn get_borders_intersection(&self) -> Option<char>

Get a border character intersection of horizontal and vertical line .

pub const fn get_borders_horizontal(&self) -> Option<char>

Get a border character horizontal .

pub const fn get_borders_vertical(&self) -> Option<char>

Get a border character vertical .

impl Theme

pub fn remove_borders_top(&mut self)

Remove a border character top .

pub fn remove_borders_bottom(&mut self)

Remove a border character bottom .

pub fn remove_borders_left(&mut self)

Remove a border character left .

pub fn remove_borders_right(&mut self)

Remove a border character right .

pub fn remove_borders_corner_top_left(&mut self)

Remove a border character top left corner .

pub fn remove_borders_corner_top_right(&mut self)

Remove a border character top right corner .

pub fn remove_borders_corner_bottom_left(&mut self)

Remove a border character bottom left corner .

pub fn remove_borders_corner_bottom_right(&mut self)

Remove a border character bottom right corner .

pub fn remove_borders_intersection_top(&mut self)

Remove a border character top intersection with a vertical line .

pub fn remove_borders_intersection_bottom(&mut self)

Remove a border character bottom intersection with a vertical line .

pub fn remove_borders_intersection_left(&mut self)

Remove a border character left intersection with a horizontal line .

pub fn remove_borders_intersection_right(&mut self)

Remove a border character right intersection with a horizontal line .

pub fn remove_borders_intersection(&mut self)

Remove a border character intersection of horizontal and vertical line .

pub fn remove_borders_horizontal(&mut self)

Remove a border character horizontal .

pub fn remove_borders_vertical(&mut self)

Remove a border character vertical .

impl Theme

pub fn set_colors_top(&mut self, color: Color)

Set a border color top .

pub fn set_colors_bottom(&mut self, color: Color)

Set a border color bottom .

pub fn set_colors_left(&mut self, color: Color)

Set a border color left .

pub fn set_colors_right(&mut self, color: Color)

Set a border color right .

pub fn set_colors_corner_top_left(&mut self, color: Color)

Set a border color top left corner .

pub fn set_colors_corner_top_right(&mut self, color: Color)

Set a border color top right corner .

pub fn set_colors_corner_bottom_left(&mut self, color: Color)

Set a border color bottom left corner .

pub fn set_colors_corner_bottom_right(&mut self, color: Color)

Set a border color bottom right corner .

pub fn set_colors_intersection_top(&mut self, color: Color)

Set a border color top intersection with a vertical line .

pub fn set_colors_intersection_bottom(&mut self, color: Color)

Set a border color bottom intersection with a vertical line .

pub fn set_colors_intersection_left(&mut self, color: Color)

Set a border color left intersection with a horizontal line .

pub fn set_colors_intersection_right(&mut self, color: Color)

Set a border color right intersection with a horizontal line .

pub fn set_colors_intersection(&mut self, color: Color)

Set a border color intersection of horizontal and vertical line .

pub fn set_colors_horizontal(&mut self, color: Color)

Set a border color horizontal .

pub fn set_colors_vertical(&mut self, color: Color)

Set a border color vertical .

impl Theme

pub fn remove_colors_top(&mut self)

Remove a border color top .

pub fn remove_colors_bottom(&mut self)

Remove a border color bottom .

pub fn remove_colors_left(&mut self)

Remove a border color left .

pub fn remove_colors_right(&mut self)

Remove a border color right .

pub fn remove_colors_corner_top_left(&mut self)

Remove a border color top left corner .

pub fn remove_colors_corner_top_right(&mut self)

Remove a border color top right corner .

pub fn remove_colors_corner_bottom_left(&mut self)

Remove a border color bottom left corner .

pub fn remove_colors_corner_bottom_right(&mut self)

Remove a border color bottom right corner .

pub fn remove_colors_intersection_top(&mut self)

Remove a border color top intersection with a vertical line .

pub fn remove_colors_intersection_bottom(&mut self)

Remove a border color bottom intersection with a vertical line .

pub fn remove_colors_intersection_left(&mut self)

Remove a border color left intersection with a horizontal line .

pub fn remove_colors_intersection_right(&mut self)

Remove a border color right intersection with a horizontal line .

pub fn remove_colors_intersection(&mut self)

Remove a border color intersection of horizontal and vertical line .

pub fn remove_colors_horizontal(&mut self)

Remove a border color horizontal .

pub fn remove_colors_vertical(&mut self)

Remove a border color vertical .

impl Theme

pub fn get_colors_top(&self) -> Option<&Color>

Get a border color top .

pub fn get_colors_bottom(&self) -> Option<&Color>

Get a border color bottom .

pub fn get_colors_left(&self) -> Option<&Color>

Get a border color left .

pub fn get_colors_right(&self) -> Option<&Color>

Get a border color right .

pub fn get_colors_corner_top_left(&self) -> Option<&Color>

Get a border color top left corner .

pub fn get_colors_corner_top_right(&self) -> Option<&Color>

Get a border color top right corner .

pub fn get_colors_corner_bottom_left(&self) -> Option<&Color>

Get a border color bottom left corner .

pub fn get_colors_corner_bottom_right(&self) -> Option<&Color>

Get a border color bottom right corner .

pub fn get_colors_intersection_top(&self) -> Option<&Color>

Get a border color top intersection with a vertical line .

pub fn get_colors_intersection_bottom(&self) -> Option<&Color>

Get a border color bottom intersection with a vertical line .

pub fn get_colors_intersection_left(&self) -> Option<&Color>

Get a border color left intersection with a horizontal line .

pub fn get_colors_intersection_right(&self) -> Option<&Color>

Get a border color right intersection with a horizontal line .

pub fn get_colors_intersection(&self) -> Option<&Color>

Get a border color intersection of horizontal and vertical line .

pub fn get_colors_horizontal(&self) -> Option<&Color>

Get a border color horizontal .

pub fn get_colors_vertical(&self) -> Option<&Color>

Get a border color vertical .