Trait tabled::settings::CellOption

pub trait CellOption<R, C> {
    // Required method
    fn change(self, records: &mut R, cfg: &mut C, entity: Entity);

    // Provided method
    fn hint_change(&self) -> Option<Entity> { ... }
}

A trait for configuring a single cell.

A cell can be targeted by [`Cell`].

[`Cell`]: crate::object::Cell

Required Methods

fn change(self, records: &mut R, cfg: &mut C, entity: Entity)

Modification function of a certail part of a grid targeted by Entity.

Provided Methods

fn hint_change(&self) -> Option<Entity>

A hint whether an [TableOption] is going to change table layout.

Return None if no changes are being done. Otherwise return:

Entity::Global - a grand layout changed.

Entity::Row - a certain row was changed.

Entity::Column - a certain column was changed.

Entity::Cell - a certain cell was changed.

By default it’s considered to be a grand change.

This methods primarily is used as an optimization, to not make unnessary calculations if they’re not needed, after using the [TableOption].