Struct tabled::style::Border

pub struct Border {
    pub top: Option<Symbol>,
    pub bottom: Option<Symbol>,
    pub left: Option<Symbol>,
    pub right: Option<Symbol>,
    pub left_top_corner: Option<Symbol>,
    pub right_top_corner: Option<Symbol>,
    pub left_bottom_corner: Option<Symbol>,
    pub right_bottom_corner: Option<Symbol>,
}

Border represents a border of a Cell.

    let table = Table::new(&data)
        .with(Style::ascii())
        .with(Modify::new(Rows::single(0)).with(Border::default().top('x')));

Fields
top: Option<Symbol>bottom: Option<Symbol>left: Option<Symbol>right: Option<Symbol>left_top_corner: Option<Symbol>right_top_corner: Option<Symbol>left_bottom_corner: Option<Symbol>right_bottom_corner: Option<Symbol>
Implementations

impl Border

Border represents a border of a Cell.

    let table = Table::new(&data)
        .with(Style::ascii())
        .with(Modify::new(Rows::single(0)).with(Border::default().top('x')));

pub fn new(
top: impl Into<Symbol>,
bottom: impl Into<Symbol>,
left: impl Into<Symbol>,
right: impl Into<Symbol>,
top_left: impl Into<Symbol>,
top_right: impl Into<Symbol>,
bottom_left: impl Into<Symbol>,
bottom_right: impl Into<Symbol>
) -> Border

This function constructs a cell borders with all sides set.

pub fn filled(c: impl Into<Symbol>) -> Border

This function constructs a cell borders with all sides’s char set to a given character. It behaives like Border::new with the same character set to each side.

pub fn top(self, c: impl Into<Symbol>) -> Border

Set a top border character.

pub fn bottom(self, c: impl Into<Symbol>) -> Border

Set a bottom border character.

pub fn left(self, c: impl Into<Symbol>) -> Border

Set a left border character.

pub fn right(self, c: impl Into<Symbol>) -> Border

Set a right border character.

pub fn top_left_corner(self, c: impl Into<Symbol>) -> Border

Set a top left intersection character.

pub fn top_right_corner(self, c: impl Into<Symbol>) -> Border

Set a top right intersection character.

pub fn bottom_left_corner(self, c: impl Into<Symbol>) -> Border

Set a bottom left intersection character.

pub fn bottom_right_corner(self, c: impl Into<Symbol>) -> Border

Set a bottom right intersection character.