Struct tabled::style::CustomStyle

pub struct CustomStyle<Top, Bottom, Left, Right, Horizontal, Vertical, Header> { /* private fields */ }

CustomStyle represents a style controlling a valid state of it.

For example. It doesn’t allow to call method CustomStyle::top_left_corner unless CustomStyle::left and CustomStyle::top is set.

Implementations

impl<T, B, L, R, IH, IV, H> CustomStyle<T, B, L, R, IH, IV, H>

pub const fn top(self, c: char) -> CustomStyle<On, B, L, R, IH, IV, H>

Sets a top border.

Any corners and intersections which were set will be overriden.

pub const fn bottom(self, c: char) -> CustomStyle<T, On, L, R, IH, IV, H>

Sets a bottom border.

Any corners and intersections which were set will be overriden.

pub const fn left(self, c: char) -> CustomStyle<T, B, On, R, IH, IV, H>

Sets a left border.

Any corners and intersections which were set will be overriden.

pub const fn right(self, c: char) -> CustomStyle<T, B, L, On, IH, IV, H>

Sets a right border.

Any corners and intersections which were set will be overriden.

pub const fn horizontal(self, c: char) -> CustomStyle<T, B, L, R, On, IV, H>

Sets a horizontal split line.

It doesn’t include a header split line. It must be set via its own method Self::header.

Any corners and intersections which were set will be overriden.

pub const fn vertical(self, c: char) -> CustomStyle<T, B, L, R, IH, On, H>

Sets a vertical split line.

Any corners and intersections which were set will be overriden.

pub const fn header(self, c: char) -> CustomStyle<T, B, L, R, IH, IV, On>

Sets a 1st horizontal split line.

Any corners and intersections which were set will be overriden.

impl<B, R, IH, IV, H> CustomStyle<On, B, On, R, IH, IV, H>

pub const fn top_left_corner(self, c: char) -> Self

Sets a top left corner.

impl<B, L, IH, IV, H> CustomStyle<On, B, L, On, IH, IV, H>

pub const fn top_right_corner(self, c: char) -> Self

Sets a top right corner.

impl<T, L, IH, IV, H> CustomStyle<T, On, L, On, IH, IV, H>

pub const fn bottom_right_corner(self, c: char) -> Self

Sets a bottom right corner.

impl<T, R, IH, IV, H> CustomStyle<T, On, On, R, IH, IV, H>

pub const fn bottom_left_corner(self, c: char) -> Self

Sets a bottom left corner.

impl<T, B, R, IV, H> CustomStyle<T, B, On, R, On, IV, H>

pub const fn left_intersection(self, c: char) -> Self

Sets a left intersection char.

impl<T, B, L, IV, H> CustomStyle<T, B, L, On, On, IV, H>

pub const fn right_intersection(self, c: char) -> Self

Sets a right intersection char.

impl<B, L, R, IH, H> CustomStyle<On, B, L, R, IH, On, H>

pub const fn top_intersection(self, c: char) -> Self

Sets a top intersection char.

impl<T, L, R, IH, H> CustomStyle<T, On, L, R, IH, On, H>

pub const fn bottom_intersection(self, c: char) -> Self

Sets a bottom intersection char.

impl<T, B, L, R, H> CustomStyle<T, B, L, R, On, On, H>

pub const fn inner_intersection(self, c: char) -> Self

Sets an inner intersection char. A char between horizontal and vertical split lines.

impl<T, B, L, R, IH> CustomStyle<T, B, L, R, IH, On, On>

pub const fn header_intersection(self, c: char) -> Self

Sets an intersection char of a 1st horizontal split line.

impl<T, B, R, IH, IV> CustomStyle<T, B, On, R, IH, IV, On>

pub const fn left_header_intersection(self, c: char) -> Self

Sets an left intersection char of a 1st horizontal split line.

impl<T, B, L, IH, IV> CustomStyle<T, B, L, On, IH, IV, On>

pub const fn right_header_intersection(self, c: char) -> Self

Sets an right intersection char of a 1st horizontal split line.

impl<B, L, R, IH, IV, H> CustomStyle<On, B, L, R, IH, IV, H>

pub const fn top_off(self) -> CustomStyle<(), B, L, R, IH, IV, H>

Removes top border.

impl<T, L, R, IH, IV, H> CustomStyle<T, On, L, R, IH, IV, H>

pub const fn bottom_off(self) -> CustomStyle<T, (), L, R, IH, IV, H>

Removes bottom border.

impl<T, B, R, IH, IV, H> CustomStyle<T, B, On, R, IH, IV, H>

pub const fn left_off(self) -> CustomStyle<T, B, (), R, IH, IV, H>

Removes left border.

impl<T, B, L, IH, IV, H> CustomStyle<T, B, L, On, IH, IV, H>

pub const fn right_off(self) -> CustomStyle<T, B, L, (), IH, IV, H>

Removes right border.

impl<T, B, L, R, IV, H> CustomStyle<T, B, L, R, On, IV, H>

pub const fn horizontal_off(self) -> CustomStyle<T, B, L, R, (), IV, H>

Removes horizontal split lines.

Not including 1st split line.

impl<T, B, L, R, IH, H> CustomStyle<T, B, L, R, IH, On, H>

pub const fn vertical_off(self) -> CustomStyle<T, B, L, R, IH, (), H>

Removes vertical split lines.

impl<T, B, L, R, IH, IV> CustomStyle<T, B, L, R, IH, IV, On>

pub const fn header_off(self) -> CustomStyle<T, B, L, R, IH, IV, ()>

Removes 1st horizontal split line.