tabled

Module tables

Module contains a list of table representatives.

Table

A default table implementation.

At it’s core it keeps data buffered. Be cautious about it.

Peek it by default.

IterTable

Just like Table but it’s API is a bit different to serve better in context where there is a memory limit.

It’s different in implementation algorithms.

From performance point of view it’s similar to Table, may be a bit slower. Test it on your specific table representation.

Peek it when you want to have a feature full table. But you have a memory conserns.

PoolTable

A table with a greater control of a layout. So you can build tables with a different layout/look easily.

Peek it when you need it.

CompactTable

A table with a limited subset of settings but it works in a no-std context. And it consumes the least amount of memory/cpu. Cause it print records one by one.

Peek it when your data contains a single line only, and you don’t need lots a features. Or you’re at no-std context.

It’s likely the fastest table in this limited context.

ExtendedTable

It’s a table which is useful for showing large amount of data. Though it’s performance is generic.

Peek it when you need it.

Structs

CompactTable

A table which consumes an IntoRecords iterator. It assumes that the content has only single line.

ExtendedTable

ExtendedTable display data in a ‘expanded display mode’ from postgresql. It may be useful for a large data sets with a lot of fields.

IterTable

A table which consumes an IntoRecords iterator.

PoolTable

PoolTable is a table which allows a greater set of possibilities for cell alignment. It’s data is not aligned in any way by default.

Table

The structure provides an interface for building a table for types that implements Tabled.

Enums

TableValue

TableValue a structure which is responsible for a PoolTable layout.