Trait tabled::Tabled

pub trait Tabled {
    fn fields(&self) -> Vec<String>;

    fn headers() -> Vec<String>;
}

Tabled a trait responsible for providing a header fields and a row fields.

It’s urgent that header len is equal to fields len.

Self::headers().len() == self.fields().len()

Required methods

fn fields(&self) -> Vec<String>

Fields method must return a list of cells.

The cells will be placed in the same row, preserving the order.

fn headers() -> Vec<String>

Headers must return a list of column names.