Module tabled::width

This module contains object which can be used to limit a cell to a given width:

Truncate cuts a cell content to limit width.

Wrap split the content via new lines in order to fit max width.

Justify sets columns width to the same value.

To set a a table width a combination of MaxWidth and MinWidth can be set.

Example

use tabled::{MaxWidth, MinWidth, Table};

let table = Table::new(&["Hello World!"])
    .with(MaxWidth::wrapping(7))
    .with(MinWidth::new(7))
    .to_string();

assert_eq!(
    table,
    concat!(
        "+-----+\n",
        "| &st |\n",
        "| r   |\n",
        "+-----+\n",
        "| Hel |\n",
        "| lo  |\n",
        "| Wor |\n",
        "| ld! |\n",
        "+-----+\n",
    )
);

Structs

Justify

Justify sets all columns widths to the set value.

Max

Max width value.

MaxWidth

MaxWidth allows you to set a max width of an object on a Table, using different strategies. It also allows you to set a MaxWidth for a whole table.

Min

Min width value.

MinWidth

MinWidth changes a content in case if it’s length is lower then the boundry.

Truncate

Truncate cut the string to a given width if its length exeeds it. Otherwise keeps the content of a cell untouched.

Wrap

Wrap wraps a string to a new line in case it exeeds the provided max boundry. Otherwise keeps the content of a cell untouched.