Struct tabled::builder::Builder

pub struct Builder { /* private fields */ }

Builder creates a Table from dynamic data set.

It useful when the amount of columns or rows is not known statically.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.push_record(["index", "measure", "value"]);
builder.push_record(["0", "weight", "0.443"]);

let table = builder.build();

println!("{}", table);

It may be useful to use FromIterator for building.

use tabled::builder::Builder;
use std::iter::FromIterator;

let data = vec![
    ["column1", "column2"],
    ["data1", "data2"],
    ["data3", "data4"],
];

let table = Builder::from_iter(data).build();

println!("{}", table);

Implementations

impl Builder

pub fn new() -> Self

Creates a Builder instance.

use tabled::builder::Builder;

let builder = Builder::new();

pub fn with_capacity(count_records: usize, count_columns: usize) -> Self

Creates a Builder instance with a given row capacity.

use tabled::builder::Builder;

let mut builder = Builder::with_capacity(2, 3);
builder.push_record((0..3).map(|i| i.to_string()));
builder.push_record(["i", "surname", "lastname"]);

pub fn from_vec(data: Vec<Vec<CellInfo<String>>>) -> Self

Creates a Builder instance.

Safety

Itโs marked unsafe to emphasize that you shall make sure that all rows bound to have the same length.

use tabled::builder::Builder;

let data = vec![];
let builder = Builder::from_vec(data);

pub fn set_empty<T>(&mut self, text: T)
where T: Into<String>,

Sets a content of cells which are created in case rows has different length.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.set_empty("undefined");
builder.push_record((0..3).map(|i| i.to_string()));
builder.push_record(["i"]);

pub fn build(self) -> Table

Build creates a Table instance.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.push_record(["i", "column1", "column2"]);
builder.push_record(["0", "value1", "value2"]);

pub fn index(self) -> IndexBuilder

Add an index to the Table.

Default index is a range 0-N where N is amount of records.

Example

use tabled::Table;

let table = Table::builder(&["Hello", "World", "!"]).index().build();

assert_eq!(
    table.to_string(),
    "+---+-------+\n\
     |   | &str  |\n\
     +---+-------+\n\
     | 0 | Hello |\n\
     +---+-------+\n\
     | 1 | World |\n\
     +---+-------+\n\
     | 2 | !     |\n\
     +---+-------+"
)

pub fn push_record<R>(&mut self, record: R)
where R: IntoIterator, R::Item: Into<String>,

Adds a row to a Table.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.push_record((0..3).map(|i| i.to_string()));
builder.push_record(["i", "surname", "lastname"]);

pub fn insert_record<R>(&mut self, index: usize, record: R)
where R: IntoIterator, R::Item: Into<String>,

Insert a row into a specific position.

Panics

Panics if index > count_rows.

pub fn clean(&mut self)

Clean removes empty columns and rows.

Example

use tabled::Table;

let mut builder = Table::builder(&["Hello", "World", ""]);
builder.clean();

let table = builder.build();

assert_eq!(
    table.to_string(),
    "+-------+\n\
     | &str  |\n\
     +-------+\n\
     | Hello |\n\
     +-------+\n\
     | World |\n\
     +-------+"
)

pub fn remove_record(&mut self, index: usize)

Removes a row with a specific position.

Index expected to be in range. Builder::count_records() < x >= 0

Panics

Panics if row_index > count_rows.

pub fn remove_column(&mut self, index: usize)

Removes a column with a specific position.

Index expected to be in range. Builder::count_columns() < x >= 0

Panics

Panics if index > count_columns.

pub fn push_column<I>(&mut self, column: I)
where I: IntoIterator, I::Item: Into<String>,

Push a column.

pub fn insert_column<I>(&mut self, index: usize, column: I)
where I: IntoIterator, I::Item: Into<String>,

Insert a column with a specific position.

In case a column is bigger then the total amount of rows it will be truncated.

Panics

Panics if index > count_columns.

pub fn clear(&mut self)

Remove all records.

pub fn count_columns(&self) -> usize

Returns an amount of columns which would be present in a built table.

pub fn count_records(&self) -> usize

Returns an amount of rows which would be present in a built table.

Notice that it does not include header if present; It returns only amount of records.