Struct tabled::builder::Builder

pub struct Builder { /* private fields */ }

Builder creates a Table from dynamic data set.

It useful when the amount of columns or rows is not known statically.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.set_header(["index", "measure", "value"]);
builder.push_record(["0", "weight", "0.443"]);

let table = builder.build();

println!("{}", table);

It may be useful to use FromIterator for building.

use tabled::builder::Builder;
use std::iter::FromIterator;

let data = vec![
    ["column1", "column2"],
    ["data1", "data2"],
    ["data3", "data4"],
];

let table = Builder::from_iter(data).build();

println!("{}", table);

Implementations

impl Builder

pub fn new() -> Self

Creates a Builder instance.

use tabled::builder::Builder;

let builder = Builder::new();

pub fn with_capacity(capacity: usize) -> Self

Creates a Builder instance with a given row capacity.

use tabled::builder::Builder;

let mut builder = Builder::with_capacity(2);
builder.push_record((0..3).map(|i| i.to_string()));
builder.push_record(["i", "surname", "lastname"]);

pub fn set_header<H, T>(&mut self, columns: H) -> &mut Selfwhere H: IntoIterator<Item = T>, T: Into<String>,

Sets a Table header.

let mut builder = Builder::default();
builder.set_header((0..3).map(|i| i.to_string()));

pub fn remove_header(&mut self) -> &mut Self

Sets off a Table header.

If not set its a nop.

use tabled::Table;

let data = [("Hello", 1u8, false), ("World", 21u8, true)];

let table = Table::builder(data).build().to_string();

assert_eq!(
    table,
    "+-------+----+-------+\n\
     | &str  | u8 | bool  |\n\
     +-------+----+-------+\n\
     | Hello | 1  | false |\n\
     +-------+----+-------+\n\
     | World | 21 | true  |\n\
     +-------+----+-------+"
);

let mut builder = Table::builder(data);
builder.remove_header();
let table = builder.build().to_string();

assert_eq!(
    table,
    "+-------+----+-------+\n\
     | Hello | 1  | false |\n\
     +-------+----+-------+\n\
     | World | 21 | true  |\n\
     +-------+----+-------+"
);

pub fn set_default_text<T>(&mut self, text: T) -> &mut Selfwhere T: Into<String>,

Sets a content of cells which are created in case rows has different length.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder
    .set_default_text("undefined")
    .set_header((0..3).map(|i| i.to_string()))
    .push_record(["i"]);

pub fn build(self) -> Table

Build creates a Table instance.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.set_header(["i", "column1", "column2"]);
builder.push_record(["0", "value1", "value2"]);

pub fn index(self) -> IndexBuilder

Add an index to the Table.

Default index is a range 0-N where N is amount of records.

Example

use tabled::Table;

let table = Table::builder(&["Hello", "World", "!"]).index().build();

assert_eq!(
    table.to_string(),
    "+---+-------+\n\
     |   | &str  |\n\
     +---+-------+\n\
     | 0 | Hello |\n\
     +---+-------+\n\
     | 1 | World |\n\
     +---+-------+\n\
     | 2 | !     |\n\
     +---+-------+"
)

pub fn push_record<R, T>(&mut self, row: R) -> &mut Selfwhere R: IntoIterator<Item = T>, T: Into<String>,

Adds a row to a Table.

use tabled::builder::Builder;

let mut builder = Builder::default();
builder.push_record((0..3).map(|i| i.to_string()));
builder.push_record(["i", "surname", "lastname"]);

pub fn insert_record<R>(&mut self, index: usize, record: R) -> boolwhere R: IntoIterator, R::Item: Into<String>,

Insert a row into a specific position.

Panics

Panics if index > count_rows.

pub fn clean(&mut self) -> &mut Self

Clean removes empty columns and rows.

Example

use tabled::Table;

let mut builder = Table::builder(&["Hello", "World", ""]);
builder.clean();

let table = builder.build();

assert_eq!(
    table.to_string(),
    "+-------+\n\
     | &str  |\n\
     +-------+\n\
     | Hello |\n\
     +-------+\n\
     | World |\n\
     +-------+"
)

pub fn hint_column_size(&mut self, size: usize) -> &mut Self

Set a column size.

If it make it lower then it was originally it is considered NOP.

pub fn count_columns(&self) -> usize

Returns an amount of columns which would be present in a built table.

pub fn count_rows(&self) -> usize

Returns an amount of rows which would be present in a built table.

pub fn has_header(&self) -> bool

Checks whether a builder contains a header set.