Struct tabled::display::ExpandedDisplay

pub struct ExpandedDisplay { /* private fields */ }

ExpandedDisplay display data in a ‘expanded display mode’ from postgress. It may be usefull for a large data sets with a lot of fields.

See ‘Examples’ in https://www.postgresql.org/docs/current/app-psql.html.

It escapes strings to resolve a multi-line ones. Because of that colors may not be rendered.

Implementations

impl ExpandedDisplay

pub fn new<T: Tabled>(iter: impl IntoIterator<Item = T>) -> Self

Creates a new instance of ExpandedDisplay

pub fn header_template(&mut self, f: fn(_: usize) -> String) -> &mut Self

Sets a line format which will be used to split records.

Default formating is “-[ RECORD {} ]-”.

At least one ‘\n’ char will be printed at the end regardless if you set it or not.

pub fn formatter(&mut self, f: impl Fn(&str) -> String + 'static) -> &mut Self

Sets a value formatter.

This method overrides others formatters like ExpandedDisplay::truncate and ExpandedDisplay::wrap.

pub fn truncate(&mut self, max: usize, tail: impl AsRef<str>) -> &mut Self

Sets max width of value. The rest will be trunceted.

pub fn wrap(&mut self, max: usize) -> &mut Self

Sets max width of value, when limit is reached next chars will be placed on the next line.