Function tabled::multiline

pub fn multiline<F>(f: F) -> Box<dyn Fn(&str) -> String> where
    F: Fn(&str) -> String + 'static, 

Multiline a helper function for changing multiline content of cell. Using this formatting applied for all rows not to a string as a whole.

use tabled::{Table, Format, multiline, Full, Modify};
let data: Vec<&'static str> = Vec::new();
let table = Table::new(&data)
    .with(Modify::new(Full).with(Format(multiline(|s| format!("{}", s)))))
    .to_string();