Module tabled::grid::config

Module contains a list of configs for varios tables/grids.

Structs

Border

Border is a representation of a cells’s borders (left, right, top, bottom, and the corners)

Borders

Borders represents a Table frame with horizontal and vertical split lines.

ColorMap

A colors structure for ColoredConfig.

ColoredConfig

A spanned configuration plus colors for cells.

CompactConfig

This structure represents a settings of a grid.

CompactMultilineConfig

A CompactConfig configuration plus vertical alignment.

EntityIterator

An iterator over cells.

EntityMap

A structure to keep information for Entity as a key.

Formatting

Formatting represent a logic of formatting of a cell.

Indent

Indent represent a filled space.

Line

A line data structure.

Sides

A structure which represents 4 box sides.

SpannedConfig

This structure represents a settings of a grid.

Enums

AlignmentHorizontal

AlignmentHorizontal represents an horizontal alignment of a cell content.

AlignmentVertical

AlignmentVertical represents an vertical alignment of a cell content.

Entity

Entity a structure which represent a set of cells.

Offset

The structure represents an offset in a text.

Type Aliases

HorizontalLine

HorizontalLine represents a horizontal border line.

Position

Position is a (row, col) position on a Grid.

VerticalLine

HorizontalLine represents a vertical border line.