Module tabled::grid::dimension

Module contains a list of implementations of Estimate and Dimension.

Structs

CompactGridDimension

A Dimension implementation which calculates exact column/row width/height.

CompleteDimension

CompleteDimension is a Dimension implementation for a Table

CompleteDimensionVecRecords

CompleteDimension is a Dimension implementation for a Table

ConstDimension

A constant size dimension or a value dimension.

PeekableDimension

PeekableDimension is a Dimension implementation for a Table

PoolTableDimension

PoolTableDimension is a dimension resolve strategy for PoolTable

SpannedGridDimension

A Dimension implementation which calculates exact column/row width/height.

SpannedVecRecordsDimension

A Dimension implementation which calculates exact column/row width/height for VecRecords.

StaticDimension

A constant dimension.

Enums

ConstSize

Const size represents either a const array values or a single value which responsible for the whole list.

DimensionPriority

A control of width/height logic for situations where we must increase some cell to align columns/row.

DimensionValue

A dimension value.

Traits

Dimension

Dimension of a Grid

Estimate

Dimension estimation of a Grid