Module tabled::grid::util::string

This module contains a different functions which are used by the Grid.

You should use it if you want to comply with how Grid.

Structs

Lines

Iterator over lines.

Functions

count_lines

Calculates a number of lines.

count_tabs

Returns a list of tabs (\t) in a string..

get_lines

Splits the string by lines.

replace_tab

Replaces tabs in a string with a given width of spaces.

string_dimension

Returns string width and count lines of a string. It’s a combination of string_width_multiline and count_lines.

string_width

Returns a string width.

string_width_multiline

Returns a max string width of a line.