Module tabled::object

Structs

Cell

Cell denotes a particular cell on a crate::Table.

Columns

Column denotes a set of cells on given columns on a crate::Table.

Combination

Combination struct used for chaning Object’s.

FirstRow

FirstRow represents the first row of a crate::Table. It’s often contains headers data.

Frame

Frame includes cells which are on the edges of each side. Therefore it’s Object implementation returns a subset of cells which are present in frame.

Full

Full represents all cells on a crate::Table

LastRow

LastRow represents the last row of a crate::Table.

Rows

Row denotes a set of cells on given rows on a crate::Table.

Segment

Segment represents a sub table of crate::Table.

Traits

Object

Object helps to locate a nessesary part of a Grid.