Module tabled::settings::height

The module contains Height structure which is responsible for a table and cell height.

Structs

CellHeightIncrease

A modification for cell/table to increase its height.

CellHeightLimit

A modification for cell/table to increase its height.

Height

Height is a abstract factory for height settings.

HeightList

A structure used to set Table height via a list of rows heights.

TableHeightIncrease

A modification of a table to increase the table height.

TableHeightLimit

A modification of a table to decrease the table height.