Module tabled::settings::location

The module contains a Location trait and implementations for it.

Example

use tabled::{
    settings::{
        location::Locator,
        object::{Columns, Object},
        Alignment, Modify, Padding,
    },
    Table, Tabled,
};

#[derive(Tabled)]
struct Reading {
    link: &'static str,
    comment: &'static str,
}

let data = [
    Reading { link: "https://www.gnu.org/software/grub/manual/multiboot/multiboot.html", comment: "todo" },
    Reading { link: "https://wiki.debian.org/initramfs", comment: "todo" },
    Reading { link: "http://jdebp.uk/FGA/efi-boot-process.html", comment: "todo,2" },
    Reading { link: "https://wiki.debian.org/UEFI", comment: "todo,2" },
];

let mut table = Table::new(data);
table.with(Padding::zero());
table.with(Modify::new(Locator::column("link")).with(Alignment::right()));
table.with(Modify::new(Locator::content("todo")).with("todo,1"));
table.with(
    Modify::new(Columns::single(1).intersect(Locator::by(|text| text.contains("todo"))))
        .with(Padding::new(4, 0, 0, 0)),
);

let output = table.to_string();

assert_eq!(
    output,
    concat!(
        "+-----------------------------------------------------------------+----------+\n",
        "|                                                             link|comment   |\n",
        "+-----------------------------------------------------------------+----------+\n",
        "|https://www.gnu.org/software/grub/manual/multiboot/multiboot.html|    todo,1|\n",
        "+-----------------------------------------------------------------+----------+\n",
        "|                                https://wiki.debian.org/initramfs|    todo,1|\n",
        "+-----------------------------------------------------------------+----------+\n",
        "|                        http://jdebp.uk/FGA/efi-boot-process.html|    todo,2|\n",
        "+-----------------------------------------------------------------+----------+\n",
        "|                                     https://wiki.debian.org/UEFI|    todo,2|\n",
        "+-----------------------------------------------------------------+----------+",
    ),
);

Structs

ByColumnName

The structure is an implementation of Location to search for a column by it’s name. A name is considered be a value in a first row.

ByCondition

The structure is an implementation of Location to search for cells with a specified condition.

ByContent

The structure is an implementation of Location to search for cells with a given content.

Locator

An abstract factory for locations, to be used to find different things on the table.

Traits

Location

Location is an interface which searches for a particular thing in the Records, and returns coordinate of the foundings if any.