Trait tabled::settings::location::Location

pub trait Location<Records> {
    type Coordinate;
    type IntoIter: IntoIterator<Item = Self::Coordinate>;

    // Required method
    fn locate(&mut self, records: &Records) -> Self::IntoIter;
}

Location is an interface which searches for a particular thing in the Records, and returns coordinate of the foundings if any.

Required Associated Types

type Coordinate

A coordinate of the finding.

type IntoIter: IntoIterator<Item = Self::Coordinate>

An iterator of the coordinates. If it’s empty it’s considered that nothing is found.

Required Methods

fn locate(&mut self, records: &Records) -> Self::IntoIter

Search for the thing in Records, returning a list of coordinates.