Trait tabled::settings::peaker::Peaker

pub trait Peaker {
    // Required methods
    fn create() -> Self;
    fn peak(&mut self, min_widths: &[usize], widths: &[usize]) -> Option<usize>;
}

A strategy of width function. It determines the order how the function is applied.

Required Methods

fn create() -> Self

Creates a new instance.

fn peak(&mut self, min_widths: &[usize], widths: &[usize]) -> Option<usize>

This function returns a column index which will be changed. Or None if no changes are necessary.