Struct tabled::settings::style::BorderColor

pub struct BorderColor<T, B, L, R> { /* private fields */ }

Border represents a border color of a Cell.

                        top border
                            |
                            V
corner top left ------> +_______+  <---- corner top left
                        |       |
left border ----------> |  cell |  <---- right border
                        |       |
corner bottom right --> +_______+  <---- corner bottom right
                            ^
                            |
                       bottom border

Example

let table = Table::new(&data)
    .with(Style::ascii())
    .modify(Rows::single(0), BorderColor::new().set_top(Color::FG_RED));

Implementations

impl BorderColor<(), (), (), ()>

pub const fn new() -> Self

Creates an empty border.

impl BorderColor<On, On, On, On>

pub const fn full( top: Color, bottom: Color, left: Color, right: Color, top_left: Color, top_right: Color, bottom_left: Color, bottom_right: Color ) -> Self

This function constructs a cell borders with all sides set.

pub fn filled(c: Color) -> Self

This function constructs a cell borders with all sides’s char set to a given color. It behaves like BorderColor::full with the same color set to each side.

impl<T, B, L, R> BorderColor<T, B, L, R>

pub fn set_top(self, c: Color) -> BorderColor<On, B, L, R>

Set a top border color.

pub fn set_bottom(self, c: Color) -> BorderColor<T, On, L, R>

Set a bottom border color.

pub fn set_left(self, c: Color) -> BorderColor<T, B, On, R>

Set a left border color.

pub fn set_right(self, c: Color) -> BorderColor<T, B, L, On>

Set a right border color.

pub fn into_inner(self) -> GridBorder<Color>

Converts a border into a general data structure.

impl<T, B, L> BorderColor<T, B, L, On>

pub fn get_right(&self) -> Color

Get a right color.

impl<T, B, R> BorderColor<T, B, On, R>

pub fn get_left(&self) -> Color

Get a left color.

impl<B, L, R> BorderColor<On, B, L, R>

pub fn get_top(&self) -> Color

Get a top color.

impl<T, L, R> BorderColor<T, On, L, R>

pub fn get_bottom(&self) -> Color

Get a bottom color.

impl<B, R> BorderColor<On, B, On, R>

pub fn set_corner_top_left(self, c: Color) -> Self

Set a top left intersection color.

pub fn get_corner_top_left(&self) -> Color

Get a top left intersection color.

impl<B, L> BorderColor<On, B, L, On>

pub fn set_corner_top_right(self, c: Color) -> Self

Set a top right intersection color.

pub fn get_corner_top_right(&self) -> Color

Get a top right intersection color.

impl<T, R> BorderColor<T, On, On, R>

pub fn set_corner_bottom_left(self, c: Color) -> Self

Set a bottom left intersection color.

pub fn get_corner_bottom_left(&self) -> Color

Get a bottom left intersection color.

impl<T, L> BorderColor<T, On, L, On>

pub fn set_corner_bottom_right(self, c: Color) -> Self

Set a bottom right intersection color.

pub fn get_corner_bottom_right(&self) -> Color

Get a bottom left intersection color.