Struct tabled::settings::themes::Theme

pub struct Theme { /* private fields */ }

A raw style data, which can be produced safely from Style.

It can be useful in order to not have a generics and be able to use it as a variable more conveniently.

Implementations

impl Theme

pub const fn from_style<T, B, L, R, H, V, const HS: usize, const VS: usize>( style: Style<T, B, L, R, H, V, HS, VS> ) -> Self

Build a theme out of a style builder.

impl Theme

pub const fn new() -> Self

Creates a new empty style.

It’s quite an analog of Style::empty

impl Theme

pub fn set_border_top(&mut self, c: char)

Set a border character top .

pub fn set_border_bottom(&mut self, c: char)

Set a border character bottom .

pub fn set_border_left(&mut self, c: char)

Set a border character left .

pub fn set_border_right(&mut self, c: char)

Set a border character right .

pub fn set_border_corner_top_left(&mut self, c: char)

Set a border character top left corner .

pub fn set_border_corner_top_right(&mut self, c: char)

Set a border character top right corner .

pub fn set_border_corner_bottom_left(&mut self, c: char)

Set a border character bottom left corner .

pub fn set_border_corner_bottom_right(&mut self, c: char)

Set a border character bottom right corner .

pub fn set_border_intersection_top(&mut self, c: char)

Set a border character top intersection with a vertical line .

pub fn set_border_intersection_bottom(&mut self, c: char)

Set a border character bottom intersection with a vertical line .

pub fn set_border_intersection_left(&mut self, c: char)

Set a border character left intersection with a horizontal line .

pub fn set_border_intersection_right(&mut self, c: char)

Set a border character right intersection with a horizontal line .

pub fn set_border_intersection(&mut self, c: char)

Set a border character intersection of horizontal and vertical line .

pub fn set_border_horizontal(&mut self, c: char)

Set a border character horizontal .

pub fn set_border_vertical(&mut self, c: char)

Set a border character vertical .

impl Theme

pub const fn get_border_top(&self) -> Option<char>

Get a border character top .

pub const fn get_border_bottom(&self) -> Option<char>

Get a border character bottom .

pub const fn get_border_left(&self) -> Option<char>

Get a border character left .

pub const fn get_border_right(&self) -> Option<char>

Get a border character right .

pub const fn get_border_corner_top_left(&self) -> Option<char>

Get a border character top left corner .

pub const fn get_border_corner_top_right(&self) -> Option<char>

Get a border character top right corner .

pub const fn get_border_corner_bottom_left(&self) -> Option<char>

Get a border character bottom left corner .

pub const fn get_border_corner_bottom_right(&self) -> Option<char>

Get a border character bottom right corner .

pub const fn get_border_intersection_top(&self) -> Option<char>

Get a border character top intersection with a vertical line .

pub const fn get_border_intersection_bottom(&self) -> Option<char>

Get a border character bottom intersection with a vertical line .

pub const fn get_border_intersection_left(&self) -> Option<char>

Get a border character left intersection with a horizontal line .

pub const fn get_border_intersection_right(&self) -> Option<char>

Get a border character right intersection with a horizontal line .

pub const fn get_border_intersection(&self) -> Option<char>

Get a border character intersection of horizontal and vertical line .

pub const fn get_border_horizontal(&self) -> Option<char>

Get a border character horizontal .

pub const fn get_border_vertical(&self) -> Option<char>

Get a border character vertical .

impl Theme

pub fn remove_border_top(&mut self)

Remove a border character top .

pub fn remove_border_bottom(&mut self)

Remove a border character bottom .

pub fn remove_border_left(&mut self)

Remove a border character left .

pub fn remove_border_right(&mut self)

Remove a border character right .

pub fn remove_border_corner_top_left(&mut self)

Remove a border character top left corner .

pub fn remove_border_corner_top_right(&mut self)

Remove a border character top right corner .

pub fn remove_border_corner_bottom_left(&mut self)

Remove a border character bottom left corner .

pub fn remove_border_corner_bottom_right(&mut self)

Remove a border character bottom right corner .

pub fn remove_border_intersection_top(&mut self)

Remove a border character top intersection with a vertical line .

pub fn remove_border_intersection_bottom(&mut self)

Remove a border character bottom intersection with a vertical line .

pub fn remove_border_intersection_left(&mut self)

Remove a border character left intersection with a horizontal line .

pub fn remove_border_intersection_right(&mut self)

Remove a border character right intersection with a horizontal line .

pub fn remove_border_intersection(&mut self)

Remove a border character intersection of horizontal and vertical line .

pub fn remove_border_horizontal(&mut self)

Remove a border character horizontal .

pub fn remove_border_vertical(&mut self)

Remove a border character vertical .

impl Theme

pub fn set_border_color_top(&mut self, color: Color)

Set a border color top .

pub fn set_border_color_bottom(&mut self, color: Color)

Set a border color bottom .

pub fn set_border_color_left(&mut self, color: Color)

Set a border color left .

pub fn set_border_color_right(&mut self, color: Color)

Set a border color right .

pub fn set_border_color_corner_top_left(&mut self, color: Color)

Set a border color top left corner .

pub fn set_border_color_corner_top_right(&mut self, color: Color)

Set a border color top right corner .

pub fn set_border_color_corner_bottom_left(&mut self, color: Color)

Set a border color bottom left corner .

pub fn set_border_color_corner_bottom_right(&mut self, color: Color)

Set a border color bottom right corner .

pub fn set_border_color_intersection_top(&mut self, color: Color)

Set a border color top intersection with a vertical line .

pub fn set_border_color_intersection_bottom(&mut self, color: Color)

Set a border color bottom intersection with a vertical line .

pub fn set_border_color_intersection_left(&mut self, color: Color)

Set a border color left intersection with a horizontal line .

pub fn set_border_color_intersection_right(&mut self, color: Color)

Set a border color right intersection with a horizontal line .

pub fn set_border_color_intersection(&mut self, color: Color)

Set a border color intersection of horizontal and vertical line .

pub fn set_border_color_horizontal(&mut self, color: Color)

Set a border color horizontal .

pub fn set_border_color_vertical(&mut self, color: Color)

Set a border color vertical .

impl Theme

pub fn remove_border_color_top(&mut self)

Remove a border color top .

pub fn remove_border_color_bottom(&mut self)

Remove a border color bottom .

pub fn remove_border_color_left(&mut self)

Remove a border color left .

pub fn remove_border_color_right(&mut self)

Remove a border color right .

pub fn remove_border_color_corner_top_left(&mut self)

Remove a border color top left corner .

pub fn remove_border_color_corner_top_right(&mut self)

Remove a border color top right corner .

pub fn remove_border_color_corner_bottom_left(&mut self)

Remove a border color bottom left corner .

pub fn remove_border_color_corner_bottom_right(&mut self)

Remove a border color bottom right corner .

pub fn remove_border_color_intersection_top(&mut self)

Remove a border color top intersection with a vertical line .

pub fn remove_border_color_intersection_bottom(&mut self)

Remove a border color bottom intersection with a vertical line .

pub fn remove_border_color_intersection_left(&mut self)

Remove a border color left intersection with a horizontal line .

pub fn remove_border_color_intersection_right(&mut self)

Remove a border color right intersection with a horizontal line .

pub fn remove_border_color_intersection(&mut self)

Remove a border color intersection of horizontal and vertical line .

pub fn remove_border_color_horizontal(&mut self)

Remove a border color horizontal .

pub fn remove_border_color_vertical(&mut self)

Remove a border color vertical .

impl Theme

pub fn get_border_color_top(&self) -> Option<&Color>

Get a border color top .

pub fn get_border_color_bottom(&self) -> Option<&Color>

Get a border color bottom .

pub fn get_border_color_left(&self) -> Option<&Color>

Get a border color left .

pub fn get_border_color_right(&self) -> Option<&Color>

Get a border color right .

pub fn get_border_color_corner_top_left(&self) -> Option<&Color>

Get a border color top left corner .

pub fn get_border_color_corner_top_right(&self) -> Option<&Color>

Get a border color top right corner .

pub fn get_border_color_corner_bottom_left(&self) -> Option<&Color>

Get a border color bottom left corner .

pub fn get_border_color_corner_bottom_right(&self) -> Option<&Color>

Get a border color bottom right corner .

pub fn get_border_color_intersection_top(&self) -> Option<&Color>

Get a border color top intersection with a vertical line .

pub fn get_border_color_intersection_bottom(&self) -> Option<&Color>

Get a border color bottom intersection with a vertical line .

pub fn get_border_color_intersection_left(&self) -> Option<&Color>

Get a border color left intersection with a horizontal line .

pub fn get_border_color_intersection_right(&self) -> Option<&Color>

Get a border color right intersection with a horizontal line .

pub fn get_border_color_intersection(&self) -> Option<&Color>

Get a border color intersection of horizontal and vertical line .

pub fn get_border_color_horizontal(&self) -> Option<&Color>

Get a border color horizontal .

pub fn get_border_color_vertical(&self) -> Option<&Color>

Get a border color vertical .

impl Theme

pub fn set_border_frame(&mut self, frame: Border<char>)

Returns an outer border of the style.

pub fn set_border_color_frame(&mut self, frame: Border<Color>)

Returns an outer border of the style.

pub fn set_border(&mut self, borders: Borders<char>)

Set borders structure.

pub fn set_border_color(&mut self, borders: Borders<Color>)

Set borders structure.

pub const fn get_border_frame(&self) -> Border<char>

Set an outer border.

pub const fn get_border_color_frame(&self) -> Border<&Color>

Set an outer border.

impl Theme

pub fn set_lines_horizontal( &mut self, lines: HashMap<usize, HorizontalLine<char>> )

Set horizontal border lines.

Example

use std::collections::HashMap;
use tabled::{Table, settings::style::{Style, HorizontalLine}, settings::themes::Theme};

let mut style = Theme::from(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, HorizontalLine::inherit(Style::extended()).into());

style.set_lines_horizontal(lines);

let data = (0..3).map(|i| ("Hello", i));
let table = Table::new(data).with(style).to_string();

assert_eq!(
    table,
    concat!(
        " ======= ===== \n",
        "  &str    i32  \n",
        "╠═══════╬═════╣\n",
        "  Hello   0    \n",
        "  Hello   1    \n",
        "  Hello   2    \n",
        " ======= ===== ",
    ),
)

pub fn set_lines_vertical(&mut self, lines: HashMap<usize, VerticalLine<char>>)

Set vertical border lines.

Example

use std::collections::HashMap;
use tabled::{
    Table,
    settings::style::{Style, HorizontalLine},
    settings::themes::Theme,
};

let mut style = Theme::from_style(Style::re_structured_text());

let mut lines = HashMap::new();
lines.insert(1, HorizontalLine::inherit(Style::extended()).into());

style.set_lines_vertical(lines);

let data = (0..3).map(|i| ("Hello", i));
let table = Table::new(data).with(style).to_string();

assert_eq!(
    table,
    concat!(
        "=======╠=====\n",
        " &str  ═ i32 \n",
        "======= =====\n",
        " Hello ═ 0   \n",
        " Hello ═ 1   \n",
        " Hello ═ 2   \n",
        "=======╣=====",
    ),
)

pub fn insert_line_vertical( &mut self, line: usize, vertical: VerticalLine<char> )

Insert a vertical line into specific column location.

pub fn insert_line_horizontal( &mut self, line: usize, horizontal: HorizontalLine<char> )

Insert a horizontal line to a specific row location.

pub fn get_line_vertical(&self, column: usize) -> Option<VerticalLine<char>>

Get a vertical line at the row if any set.

pub fn get_line_horizontal(&self, row: usize) -> Option<HorizontalLine<char>>

Get a horizontal line at the row if any set.

impl Theme

pub fn reverse_rows(&mut self, reverse: bool)

Reverse rows.

pub fn reverse_columns(&mut self, reverse: bool)

Reverse columns.

pub fn set_footer(&mut self, footer: bool)

Set a footer.

Copy columns names to an apposite side of a table.

pub fn align_columns(&mut self, position: Alignment)

Set column alignment