Module tabled::tables

Module contains a list of table representatives.

Table

A default table implementation.

IterTable

Just like Table but it’s API is a bit different to serve better in context where there is a memory limit.

ExtendedTable

It’s a table which is useful for large amount of data.

PoolTable

A table with a greather controll of a layout.

Structs

CompactTable

A table which consumes an IntoRecords iterator. It assumes that the content has only single line.

ExtendedTable

ExtendedTable display data in a ‘expanded display mode’ from postgresql. It may be useful for a large data sets with a lot of fields.

IterTable

A table which consumes an IntoRecords iterator.

PoolTable

PoolTable is a table which allows a greater set of possibilities for cell alignment. It’s data is not aligned in any way by default.

Table

The structure provides an interface for building a table for types that implements Tabled.

Enums

TableValue

TableValue a structure which is responsible for a PoolTable layout.