Trait tabled::Tabled

pub trait Tabled {
    const LENGTH: usize;

    // Required methods
    fn fields(&self) -> Vec<Cow<'_, str>>;
    fn headers() -> Vec<Cow<'static, str>>;
}

Tabled a trait responsible for providing a header fields and a row fields.

It’s urgent that header len is equal to fields len.

Self::headers().len() == self.fields().len()

Required Associated Constants

const LENGTH: usize

A length of fields and headers, which must be the same.

Required Methods

fn fields(&self) -> Vec<Cow<'_, str>>

Fields method must return a list of cells.

The cells will be placed in the same row, preserving the order.

fn headers() -> Vec<Cow<'static, str>>

Headers must return a list of column names.